refactor(pixelflow-search): encapsulate the guide behind its contract; prune the dead remainder
The Guide's contractual role (docs/plans/2026-08-17-cost-model-domain.md,
J10): given the e-graph's current state and a set of candidate rule
applications, produce per-candidate move-ordering scores for guided
saturation — trained later, purely supervised, on hindsight provenance
labels from egraph::labeler.

Everything that implements that contract moves behind a new
`nnue::guide` module whose only public surface is the `SaturationGuide`
trait and its constructor: `guide/accumulator.rs` (GraphAccumulator and
its 18 mutators), `guide/scoring.rs` (mask MLP, bilinear scoring,
encode_rule_from_arena). It is KEPT, not deleted-and-rewritten: JP's
instruction was that if we're going to want it later, don't delete and
rewrite it — this surface is inert until Phase 3 of the guided-saturation
redesign (2026-07-07) starts training it. Characterization tests moved
with the code they characterize, unchanged. 1,162 lines moved behind the
boundary (factored.rs -1,996; guide/ +1,162 across accumulator.rs,
scoring.rs, mod.rs).

What was actually deleted has zero callers and zero roadmap role: the
`Edge` struct, `ExprNnue::from_factored`, and the legacy `RuleFeatures` /
`set_rule` / `rule_mlp_*` / `encode_rule` / `encode_all_rules` /
`encode_all_rules_from_templates` sub-tier, superseded by the arena-based
`encode_rule_from_arena` this same encapsulation keeps live. That's the
other 834 lines of factored.rs's net reduction. `apply_unified_sgd` and
`UnifiedGradients` (pixelflow-pipeline/src/training/unified_backward.rs)
drop every `d_mask_*`/`d_graph_*` field as a forced consequence — this is
the confirmed decay-of-dead-weights defect: Phase-3 weights were
randomly initialized then weight-decayed every run with no opposing
gradient, and now there is no gradient struct for them to decay through.

Live-weights-format consequence (kills P1(b), the never-bumped-on-
feature-change checkpoint schema): the on-disk format is now live-only —
extraction head + shared backbone, no mask/graph/rule noise mixed in.
Magic bumped TRIE -> TRIF. A pre-existing TRIE file is refused loudly
(`ExprNnue::load`/`from_bytes` returns an `InvalidData` error naming both
magics; every caller either panics with the diff or bubbles the error up
per NO SILENT FAILURES) — verified by hand against a fabricated
TRIE-magic file this round, and by `judge_weights_load.rs`'s round-trip
test in CI. Old files require retrain, not migration: deliberately no
format-evolution ceremony for a scratch artifact that's cheaper to
remint than to carry compatibility code for.

Also folds in the egraph-crate half of the saturation-loop dedup (J11):
`EGraph::saturate_with_limits` becomes the one rewrite-until-budget-
exhausted loop and returns `SaturationStats`; `egraph::saturate`'s
`saturate_with_full_budget` calls it instead of re-deciding the same
stopping conditions in a second, hand-rolled outer loop — the exact
duplicate-loop drift the domain-model doc calls out by name.
