+++ pixelflow-pipeline/tests/judge_weights_load.rs
@@ -1,6 +1,7 @@
 //! Verifies the Judge extraction-head weights round-trip through the
-//! production serializer: `ExprNnue::save` (which writes the current "TRIE"
-//! format) followed by `ExprNnue::from_bytes` (which only accepts TRIE).
+//! production serializer: `ExprNnue::save` (which writes the current "TRIF"
+//! live-only format) followed by `ExprNnue::from_bytes` (which only accepts
+//! TRIF).
 //!
 //! Phase 2 of docs/plans/2026-07-07-guided-saturation-redesign.md: the
 //! shipped state before the retrain was NO weights (old TRIC-format file
@@ -13,11 +14,17 @@
 //! `.bin`, so it passes on a fresh checkout and in CI. Whether an *actually
 //! retrained* model beats the latency prior is the job of the
 //! `bench_extraction_3way` benchmark, not a unit test.
+//!
+//! TRIE (this test's name before 2026-08-17) also serialized the saturation
+//! head's untrained parameters; TRIF is live-only (extraction head + shared
+//! backbone), per docs/plans/2026-08-17-cost-model-domain.md item 2 — TRIE
+//! files require retrain, not migration (deliberately no format-evolution
+//! ceremony for a scratch artifact cheaper to remint than to migrate).
 
 use pixelflow_search::nnue::factored::ExprNnue;
 
 #[test]
-fn judge_weights_round_trip_via_trie() {
+fn judge_weights_round_trip_via_trif() {
     // A latency-prior-initialized model is well-formed and non-zero — the
     // same code path `bootstrap_extraction_head` starts from before training.
     let model = ExprNnue::new_with_latency_prior(0xF00D_2026);
@@ -33,10 +40,10 @@ fn judge_weights_round_trip_via_trie() {
         std::fs::read(&path).unwrap_or_else(|e| panic!("failed to read {}: {e}", path.display()));
     let _ = std::fs::remove_file(&path);
 
-    // The magic must be TRIE (not a stale TRID/TRIC), and the loader must accept it.
-    assert_eq!(&bytes[0..4], b"TRIE", "saved file is not TRIE-format");
+    // The magic must be TRIF (not a stale TRIE/TRID/TRIC), and the loader must accept it.
+    assert_eq!(&bytes[0..4], b"TRIF", "saved file is not TRIF-format");
     let loaded = ExprNnue::from_bytes(&bytes)
-        .unwrap_or_else(|e| panic!("ExprNnue::from_bytes rejected a freshly-saved TRIE file: {e}"));
+        .unwrap_or_else(|e| panic!("ExprNnue::from_bytes rejected a freshly-saved TRIF file: {e}"));
 
     // Beyond "magic matched": a model whose embeddings are all zero/non-finite
     // would still parse as valid TRIE but carry no signal. Guard against a
