+++ pixelflow-search/src/nnue/factored.rs
@@ -59,21 +59,15 @@ pub const K: usize = 32;
 /// For the edge tower (`w1`, [`INPUT_DIM`]) these four slots carry the
 /// variance histogram (const / frame / scanline / pixel fractions) — see
 /// [`EdgeAccumulator::extraction_input`], the single place that builds them.
-/// For the graph tower (`graph_w1`, [`GRAPH_INPUT_DIM`]) they carry
-/// log2-compressed search-resource scalars (edge_count, node_count,
-/// node_budget, epoch_budget).
+/// `nnue::guide`'s graph tower reuses this constant for the same reason
+/// (log2-compressed search-resource scalars), so it stays `pub` rather than
+/// `pub(crate)`.
 pub const SCALAR_FEATURE_COUNT: usize = 4;
 
 /// Total input dimension to the hidden layer:
 /// 4K (dual accumulator: 2K flat + 2K depth-encoded) + 4 scalars
 /// (the variance histogram — see [`EdgeAccumulator::extraction_input`]).
 pub const INPUT_DIM: usize = 4 * K + SCALAR_FEATURE_COUNT;
-
-/// Graph accumulator dimension: marginals (2K) + 1-hop VSA binding (K) + 2-hop VSA binding (K).
-pub const GRAPH_ACC_DIM: usize = 4 * K; // 128
-
-/// Graph backbone input: 4K + 4 scalars (edge_count, node_count, node_budget, epoch_budget).
-pub const GRAPH_INPUT_DIM: usize = GRAPH_ACC_DIM + SCALAR_FEATURE_COUNT; // 132
 
 /// Maximum arity for child-index encoding.
 /// Effective depth = `depth * MAX_ARITY + child_index`, where child_index ∈ [0, MAX_ARITY).
@@ -89,110 +83,17 @@ pub const MAX_DEPTH: usize = 192;
 pub const HIDDEN_DIM: usize = 64;
 
 // ============================================================================
-// Unified Mask Architecture Constants
+// Shared Embedding Constants
 // ============================================================================
 
-/// Embedding dimension for expr/rule factorization in the unified mask architecture.
+/// Embedding dimension for the expr/rule/graph embedding space. Shared by the
+/// live extraction head (`expr_proj`, `value_mlp`) and by `nnue::guide`'s
+/// saturation head (mask MLP, rule projection) — one embedding space, two
+/// downstream heads.
 pub const EMBED_DIM: usize = 32;
 
-/// Hidden dimension for private MLPs (value, mask, rule).
+/// Hidden dimension for private per-head MLPs.
 pub const MLP_HIDDEN: usize = 16;
-
-/// Rule feature dimension (hand-crafted features describing each rule).
-pub const RULE_FEATURE_DIM: usize = 8;
-
-/// Maximum rules supported in the unified mask architecture.
-/// Designed to scale to 1000+ rules.
-pub const MASK_MAX_RULES: usize = 1024;
-
-/// Concatenated rule features: [z_LHS | z_RHS | z_LHS-z_RHS | z_LHS*z_RHS] (4 × EMBED_DIM).
-/// Used when encoding rules from their LHS/RHS expression templates.
-pub const RULE_CONCAT_DIM: usize = 4 * EMBED_DIM;
-
-/// Mask MLP input dimension: expr_embed directly (EMBED_DIM = 32 dims).
-/// value_pred was removed — it is a deterministic function of expr_embed and adds zero information.
-pub const MASK_INPUT_DIM: usize = EMBED_DIM;
-
-// NOTE: SEARCH_INPUT_DIM removed - mask IS the policy.
-// See plan: "Idea 4B: Mask IS the search/policy ✅ CHOSEN"
-
-// ============================================================================
-// Rule Features
-// ============================================================================
-
-/// Hand-crafted features describing each rule.
-///
-/// These features are mostly static (computed once when rules are defined)
-/// and allow the Rule MLP to generalize across rules without learning
-/// individual embeddings for each rule.
-///
-/// # Features (RULE_FEATURE_DIM = 8)
-///
-/// 1. `category`: Rule type (algebraic=0, peephole=0.25, domain=0.5, cross-cutting=0.75)
-/// 2. `lhs_nodes`: Pattern complexity (normalized by 10)
-/// 3. `typical_depth_delta`: Usually -1, 0, or 1
-/// 4. `commutative`: Does rule exploit commutativity? (0 or 1)
-/// 5. `associative`: Does rule exploit associativity? (0 or 1)
-/// 6. `creates_sharing`: Does rule typically enable CSE? (0 or 1)
-/// 7. `historical_match_rate`: Running average [0, 1]
-/// 8. `expensive_op_related`: Touches div/sqrt/transcendental? (0 or 1)
-#[derive(Clone)]
-pub struct RuleFeatures {
-    /// Features for each rule: [rule_idx][feature_dim]
-    pub features: [[f32; RULE_FEATURE_DIM]; MASK_MAX_RULES],
-}
-
-impl Default for RuleFeatures {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl RuleFeatures {
-    /// Create zero-initialized rule features.
-    #[must_use]
-    pub fn new() -> Self {
-        Self {
-            features: [[0.0; RULE_FEATURE_DIM]; MASK_MAX_RULES],
-        }
-    }
-
-    /// Get features for a specific rule.
-    #[must_use]
-    pub fn get(&self, rule_idx: usize) -> &[f32; RULE_FEATURE_DIM] {
-        &self.features[rule_idx]
-    }
-
-    /// Set features for a specific rule.
-    pub fn set(&mut self, rule_idx: usize, features: [f32; RULE_FEATURE_DIM]) {
-        self.features[rule_idx] = features;
-    }
-
-    /// Set feature by name for easier initialization.
-    pub fn set_rule(
-        &mut self,
-        rule_idx: usize,
-        category: f32,
-        lhs_nodes: usize,
-        depth_delta: i8,
-        commutative: bool,
-        associative: bool,
-        creates_sharing: bool,
-        match_rate: f32,
-        expensive_op: bool,
-    ) {
-        self.features[rule_idx] = [
-            category,
-            lhs_nodes as f32 / 10.0,
-            depth_delta as f32,
-            if commutative { 1.0 } else { 0.0 },
-            if associative { 1.0 } else { 0.0 },
-            if creates_sharing { 1.0 } else { 0.0 },
-            match_rate.clamp(0.0, 1.0),
-            if expensive_op { 1.0 } else { 0.0 },
-        ];
-    }
-}
 
 // ============================================================================
 // Rule Templates (LHS/RHS Expression Templates)
@@ -654,32 +555,6 @@ const fn const_cos(x: f64) -> f64 {
 #[inline]
 pub fn depth_pe(depth: u32) -> &'static [f32; K] {
     &DEPTH_PE[depth.min((MAX_DEPTH - 1) as u32) as usize]
-}
-
-/// Cyclic rotation by `amount % K` positions (generalised VSA permutation).
-///
-/// `shift_by(emb, 0)` is the identity, `shift_by(emb, 1)` is the original
-/// `shift1`, and higher amounts encode hierarchical depth in VSA bindings:
-/// `parent ⊙ shift_by(child, depth)` produces a distinct binding per depth
-/// level, so `Add(Add(X,Y),Z)` and `Add(X,Add(Y,Z))` yield different
-/// accumulators.
-#[inline]
-fn shift_by(emb: &[f32; K], amount: usize) -> [f32; K] {
-    let amount = amount % K;
-    let mut out = [0.0f32; K];
-    for i in 0..K {
-        out[i] = emb[(i + amount) % K];
-    }
-    out
-}
-
-/// Cyclic shift by 1 position (VSA permutation for breaking commutativity).
-///
-/// Used by `GraphAccumulator` to ensure `parent ⊙ shift₁(child)` produces a
-/// different binding vector than `child ⊙ shift₁(parent)`, i.e. `Mul→Add ≠ Add→Mul`.
-#[inline]
-fn shift1(emb: &[f32; K]) -> [f32; K] {
-    shift_by(emb, 1)
 }
 
 // ============================================================================
@@ -1293,369 +1168,27 @@ impl CostDag for ChoicesCostDag<'_> {
 }
 
 // ============================================================================
-// Graph Accumulator (VSA encoding of e-graph state for saturation head)
+// Extraction (Value) Head
 // ============================================================================
 
-/// VSA accumulator for e-graph state (rebuilt each epoch).
-///
-/// Three-section encoding captures both marginal and joint op distributions:
-///
-/// | Section | Dim | Operation | Signal |
-/// |---------|-----|-----------|--------|
-/// | `[0..K]` | K | `Σ E[parent]` | Marginal: which ops appear as parents |
-/// | `[K..2K]` | K | `Σ E[child]` | Marginal: which ops appear as children |
-/// | `[2K..3K]` | K | `Σ E[parent] ⊙ shift₁(E[child])` | **1-hop VSA binding**: which ops are connected |
-/// | `[3K..4K]` | K | `Σ E[gp] ⊙ shift₁(E[par]) ⊙ shift²(E[child])` | **2-hop VSA binding**: 3-node path patterns |
-///
-/// The 1-hop binding section uses element-wise Hadamard product with a cyclic shift to
-/// break commutativity (`Mul→Add ≠ Add→Mul`). This captures the **joint**
-/// distribution of parent-child pairs — strictly more informative than marginals
-/// alone.
-///
-/// The 2-hop binding section extends this to 3-node paths (grandparent→parent→child),
-/// capturing patterns like "Mul feeding Add feeding Sqrt" which is exactly what
-/// rewrite rules match on. This turns the accumulator from a 0-round GNN into
-/// a 1-round GNN.
-///
-/// The downstream backbone learns to decode the bundled representation.
-///
-/// Shares `OpEmbeddings` with the extraction head — same learned op embeddings,
-/// different downstream pathway.
-#[derive(Clone)]
-pub struct GraphAccumulator {
-    /// `[0..K]`:    marginal parent sum     `Σ E[parent]`
-    /// `[K..2K]`:   marginal child sum      `Σ E[child]`
-    /// `[2K..3K]`:  1-hop VSA binding sum   `Σ E[parent] ⊙ shift_by(E[child], depth)`
-    /// `[3K..4K]`:  2-hop VSA binding sum   `Σ E[gp] ⊙ shift₁(E[par]) ⊙ shift²(E[child])`
-    pub values: [f32; GRAPH_ACC_DIM],
-    /// Number of edges added to the accumulator.
-    pub edge_count: u32,
-    /// Number of nodes (ops + leaves) in the graph.
-    pub node_count: u32,
-    /// E-graph node budget for this trajectory (how many nodes the saturator may create).
-    pub node_budget: u32,
-    /// Epoch budget for this trajectory (max saturation epochs).
-    pub epoch_budget: u32,
-}
-
-impl Default for GraphAccumulator {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl GraphAccumulator {
-    /// Create a zero-initialized graph accumulator.
-    #[must_use]
-    pub fn new() -> Self {
-        Self {
-            values: [0.0; GRAPH_ACC_DIM],
-            edge_count: 0,
-            node_count: 0,
-            node_budget: 0,
-            epoch_budget: 0,
-        }
-    }
-
-    /// Reset to zero state.
-    ///
-    /// Budget fields are intentionally NOT reset — they are trajectory-level
-    /// properties that should persist across epoch rebuilds.
-    pub fn reset(&mut self) {
-        self.values = [0.0; GRAPH_ACC_DIM];
-        self.edge_count = 0;
-        self.node_count = 0;
-    }
-
-    /// Add a single edge with depth-aware VSA encoding.
-    ///
-    /// Updates all three sections: marginal parent, marginal child, and
-    /// VSA binding (`E[parent] ⊙ shift_by(E[child], depth % K)`).
-    ///
-    /// At `depth == 0` the shift is the identity (root edges), `depth == 1`
-    /// shifts by 1 (matching the original `shift1` behavior), `depth == 2`
-    /// shifts by 2, etc. This encodes hierarchical position into the binding
-    /// without any extra parameters.
-    #[inline]
-    pub fn add_edge_at_depth(
-        &mut self,
-        emb: &OpEmbeddings,
-        parent_op: OpKind,
-        child_op: OpKind,
-        depth: usize,
-    ) {
-        let p = emb.get(parent_op);
-        let c = emb.get(child_op);
-        let c_shifted = shift_by(c, depth);
-        for i in 0..K {
-            self.values[i] += p[i]; // marginal parent
-            self.values[K + i] += c[i]; // marginal child
-            self.values[2 * K + i] += p[i] * c_shifted[i]; // VSA binding
-        }
-        self.edge_count += 1;
-    }
-
-    /// Add a single edge with VSA encoding (backward-compatible, depth = 1).
-    ///
-    /// Equivalent to `add_edge_at_depth(emb, parent_op, child_op, 1)`.
-    /// Preserves the original `shift1` behavior for callers that do not
-    /// track depth.
-    #[inline]
-    pub fn add_edge(&mut self, emb: &OpEmbeddings, parent_op: OpKind, child_op: OpKind) {
-        self.add_edge_at_depth(emb, parent_op, child_op, 1);
-    }
-
-    /// Add a leaf node (Var/Const) — no edges, just increment node count.
-    pub fn add_leaf(&mut self) {
-        self.node_count += 1;
-    }
-
-    /// Add an Op node and all its edges to children, with depth-aware VSA.
-    ///
-    /// Emits one `add_edge_at_depth` per child and increments `node_count`
-    /// once.  The `depth` parameter is the depth of `op` in the expression
-    /// tree (0 = root).
-    pub fn add_op_node_at_depth(
-        &mut self,
-        emb: &OpEmbeddings,
-        op: OpKind,
-        child_ops: &[OpKind],
-        depth: usize,
-    ) {
-        for &child_op in child_ops {
-            self.add_edge_at_depth(emb, op, child_op, depth);
-        }
-        self.node_count += 1;
-    }
-
-    /// Add an Op node and all its edges to children (backward-compatible, depth = 1).
-    ///
-    /// Equivalent to `add_op_node_at_depth(emb, op, child_ops, 1)`.
-    pub fn add_op_node(&mut self, emb: &OpEmbeddings, op: OpKind, child_ops: &[OpKind]) {
-        self.add_op_node_at_depth(emb, op, child_ops, 1);
-    }
-
-    // ========== Incremental Removal (inverse of addition) ==========
-
-    /// Remove a single edge with depth-aware VSA encoding — the exact
-    /// inverse of [`add_edge_at_depth`].
-    ///
-    /// Subtracts (instead of adds) the parent, child, and VSA binding
-    /// contributions from each section. Decrements `edge_count` via
-    /// saturating subtraction so underflow clamps to zero rather than
-    /// wrapping.
-    ///
-    /// # Contract
-    ///
-    /// Callers must only remove edges that were previously added at the
-    /// same `depth`. Removing an edge that was never added will corrupt
-    /// the accumulator values (negative contributions) and is a logic
-    /// error.
-    #[inline]
-    pub fn remove_edge_at_depth(
-        &mut self,
-        emb: &OpEmbeddings,
-        parent_op: OpKind,
-        child_op: OpKind,
-        depth: usize,
-    ) {
-        let p = emb.get(parent_op);
-        let c = emb.get(child_op);
-        let c_shifted = shift_by(c, depth);
-        for i in 0..K {
-            self.values[i] -= p[i]; // marginal parent
-            self.values[K + i] -= c[i]; // marginal child
-            self.values[2 * K + i] -= p[i] * c_shifted[i]; // VSA binding
-        }
-        self.edge_count = self.edge_count.saturating_sub(1);
-    }
-
-    /// Remove a single edge (backward-compatible, depth = 1) — the exact
-    /// inverse of [`add_edge`].
-    ///
-    /// Equivalent to `remove_edge_at_depth(emb, parent_op, child_op, 1)`.
-    #[inline]
-    pub fn remove_edge(&mut self, emb: &OpEmbeddings, parent_op: OpKind, child_op: OpKind) {
-        self.remove_edge_at_depth(emb, parent_op, child_op, 1);
-    }
-
-    /// Remove an Op node and all its edges to children with depth-aware
-    /// VSA — the exact inverse of [`add_op_node_at_depth`].
-    ///
-    /// Calls [`remove_edge_at_depth`] for each child and decrements
-    /// `node_count`.
-    pub fn remove_op_node_at_depth(
-        &mut self,
-        emb: &OpEmbeddings,
-        op: OpKind,
-        child_ops: &[OpKind],
-        depth: usize,
-    ) {
-        for &child_op in child_ops {
-            self.remove_edge_at_depth(emb, op, child_op, depth);
-        }
-        self.node_count = self.node_count.saturating_sub(1);
-    }
-
-    /// Remove an Op node and all its edges (backward-compatible, depth = 1)
-    /// — the exact inverse of [`add_op_node`].
-    ///
-    /// Equivalent to `remove_op_node_at_depth(emb, op, child_ops, 1)`.
-    pub fn remove_op_node(&mut self, emb: &OpEmbeddings, op: OpKind, child_ops: &[OpKind]) {
-        self.remove_op_node_at_depth(emb, op, child_ops, 1);
-    }
-
-    /// Remove a leaf node — the exact inverse of [`add_leaf`].
-    ///
-    /// Decrements `node_count` only (leaves contribute no edges).
-    pub fn remove_leaf(&mut self) {
-        self.node_count = self.node_count.saturating_sub(1);
-    }
-
-    // ========== 2-hop Message Passing (1-round GNN) ==========
-
-    /// Add a 2-hop (grandparent→parent→child) binding to the `[3K..4K]` section.
-    ///
-    /// Encodes 3-node path patterns like "Mul feeding Add feeding Sqrt" using
-    /// the VSA triple product `E[grandparent] ⊙ shift₁(E[parent]) ⊙ shift²(E[child])`.
-    /// The shift amounts break commutativity: `A→B→C` produces a different
-    /// binding than any permutation of {A, B, C}.
-    ///
-    /// Does NOT modify the `[0..3K]` sections or `edge_count`/`node_count` —
-    /// those are maintained by `add_edge*` / `add_op_node*`.
-    #[inline]
-    pub fn add_2hop_edge(
-        &mut self,
-        emb: &OpEmbeddings,
-        grandparent_op: OpKind,
-        parent_op: OpKind,
-        child_op: OpKind,
-    ) {
-        let gp = emb.get(grandparent_op);
-        let p = shift1(emb.get(parent_op));
-        let c = shift_by(emb.get(child_op), 2);
-        for i in 0..K {
-            self.values[3 * K + i] += gp[i] * p[i] * c[i];
-        }
-    }
-
-    /// Remove a 2-hop (grandparent→parent→child) binding — the exact inverse
-    /// of [`add_2hop_edge`].
-    ///
-    /// Subtracts the triple-product contribution from the `[3K..4K]` section.
-    ///
-    /// # Contract
-    ///
-    /// Callers must only remove 2-hop edges that were previously added with
-    /// the same (grandparent, parent, child) triple. Removing a path that was
-    /// never added will corrupt the accumulator values and is a logic error.
-    #[inline]
-    pub fn remove_2hop_edge(
-        &mut self,
-        emb: &OpEmbeddings,
-        grandparent_op: OpKind,
-        parent_op: OpKind,
-        child_op: OpKind,
-    ) {
-        let gp = emb.get(grandparent_op);
-        let p = shift1(emb.get(parent_op));
-        let c = shift_by(emb.get(child_op), 2);
-        for i in 0..K {
-            self.values[3 * K + i] -= gp[i] * p[i] * c[i];
-        }
-    }
-
-    /// Return a copy with each of the four sections independently L2-normalized.
-    ///
-    /// Raw sums grow proportionally to edge count, so a 200-edge graph has
-    /// values ~20x larger than a 10-edge graph.  Normalizing makes the
-    /// embedding scale-invariant: small rewrites on large graphs become visible
-    /// instead of being swamped by magnitude.
-    ///
-    /// Each section is normalized independently because they represent different
-    /// quantities with different natural scales:
-    /// - `[0..K]`    marginal parent sums
-    /// - `[K..2K]`   marginal child sums
-    /// - `[2K..3K]`  1-hop VSA binding sums
-    /// - `[3K..4K]`  2-hop VSA binding sums
-    ///
-    /// Scalar fields (`edge_count`, `node_count`, etc.) are copied as-is.
-    ///
-    /// A zero-norm section (no edges accumulated) is left as all-zeros rather
-    /// than producing NaN/Inf.
-    #[must_use]
-    pub fn normalized(&self) -> Self {
-        let mut out = self.clone();
-        out.normalize_in_place();
-        out
-    }
-
-    /// L2-normalize each of the four sections in place.
-    ///
-    /// See [`normalized`](Self::normalized) for rationale.
-    pub fn normalize_in_place(&mut self) {
-        l2_normalize_section(&mut self.values, 0, K);
-        l2_normalize_section(&mut self.values, K, 2 * K);
-        l2_normalize_section(&mut self.values, 2 * K, 3 * K);
-        l2_normalize_section(&mut self.values, 3 * K, 4 * K);
-    }
-}
-
-/// L2-normalize a contiguous slice `values[start..end]` in place.
-///
-/// If the section norm is zero (or negligibly small), it is left untouched
-/// to avoid division by zero.
-fn l2_normalize_section(values: &mut [f32], start: usize, end: usize) {
-    let mut sum_sq: f32 = 0.0;
-    for i in start..end {
-        sum_sq += values[i] * values[i];
-    }
-    let norm = sqrtf(sum_sq);
-    // Guard: skip normalization for zero/near-zero sections to avoid NaN/Inf.
-    if norm < 1e-12 {
-        return;
-    }
-    let inv_norm = 1.0 / norm;
-    for i in start..end {
-        values[i] *= inv_norm;
-    }
-}
-
-// ============================================================================
-// Structural Hashing
-// ============================================================================
-
-// ============================================================================
-// Edge Extraction Utilities
-// ============================================================================
-
-/// An edge in the expression tree: (parent_op, child_op).
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
-pub struct Edge {
-    /// The parent operation type.
-    pub parent: OpKind,
-    /// The child operation type.
-    pub child: OpKind,
-}
-
-// ============================================================================
-// Dual-Head NNUE (AlphaZero-style)
-// ============================================================================
-
-/// ExprNnue: shared backbone with one extraction head (Value MLP) and one saturation head (bilinear mask).
+/// `ExprNnue`: the shared backbone plus the extraction (value) head.
 ///
 /// ## Architecture
 ///
 /// ```text
-/// expr → OpEmbeddings → EdgeAccumulator → hidden [64] → expr_proj → expr_embed [32]
-///                                                            ├─→ value_mlp → cost (extraction head)
-///                                                            └─→ [embed, cost] → mask_mlp → bilinear → score (saturation head)
+/// expr → OpEmbeddings → EdgeAccumulator → hidden [64] → trunk → expr_proj → expr_embed [32] → value_mlp → cost
 /// ```
 ///
-/// **Extraction head**: `expr_embed → value_mlp (32→16→1)` predicts log-nanosecond cost.
-/// **Saturation head**: `[expr_embed, value_pred] → mask_mlp → bilinear(mask_features, rule_embed)` scores rules.
+/// `expr_embed → value_mlp (32→16→1)` predicts log-nanosecond cost.
 ///
-/// Rule embeddings come from LHS/RHS templates via `rule_proj`, not from learned per-rule embeddings.
+/// This struct is deliberately extraction-only: it is the live NNUE
+/// checkpoint format ("TRIF" — see [`Self::save`]), and every param in it is
+/// trained by `bootstrap_extraction_head`. The saturation (mask/policy) head
+/// that used to share this struct — inert, zero non-test callers, Phase-3
+/// gated — now lives behind [`crate::nnue::guide::SaturationGuide`], which
+/// holds its own weights and reads this struct's shared trunk (`trunk_w`,
+/// `trunk_b`), embeddings, and `expr_proj`/`forward_expr_only` by reference
+/// rather than duplicating them.
 #[derive(Clone)]
 pub struct ExprNnue {
     // ========== SHARED (Expression Backbone) ==========
@@ -1669,17 +1202,14 @@ pub struct ExprNnue {
     pub b1: [f32; HIDDEN_DIM],
 
     /// Shared trunk weights: [HIDDEN_DIM][HIDDEN_DIM] (64 x 64 = 4,096 params).
-    /// Both edge tower and graph tower outputs pass through this layer before
-    /// reaching their task-specific projection heads.
-    /// This is the "deep conceptual representation" shared between extraction and saturation.
+    /// Read by `nnue::guide`'s graph tower too (see `apply_trunk`), so this
+    /// remains "the deep conceptual representation" shared across both heads
+    /// even though the mask/graph tower's own weights moved out.
     pub trunk_w: [[f32; HIDDEN_DIM]; HIDDEN_DIM],
     /// Shared trunk biases: [HIDDEN_DIM] (64 params).
     pub trunk_b: [f32; HIDDEN_DIM],
 
-    // ========== UNIFIED MASK ARCHITECTURE ==========
-    // These fields support the new bilinear expr-rule interaction model
-    // that scales to 1000+ rules.
-    /// Projects backbone hidden (64) to shared expr embedding (EMBED_DIM=32).
+    /// Projects backbone hidden (64) to the shared expr embedding (`EMBED_DIM`=32).
     /// Weights: [HIDDEN_DIM x EMBED_DIM]
     pub expr_proj_w: [[f32; EMBED_DIM]; HIDDEN_DIM],
     /// Expr projection bias: [EMBED_DIM]
@@ -1693,57 +1223,6 @@ pub struct ExprNnue {
     pub value_mlp_w2: [f32; MLP_HIDDEN],
     /// Value MLP layer 2 bias
     pub value_mlp_b2: f32,
-
-    /// Mask MLP layer 1 weights: expr_embed (32) → hidden (16).
-    /// value_pred was removed from the input — it is a deterministic function
-    /// of expr_embed and adds zero information (see `MASK_INPUT_DIM`).
-    pub mask_mlp_w1: [[f32; MLP_HIDDEN]; MASK_INPUT_DIM],
-    /// Mask MLP layer 1 bias
-    pub mask_mlp_b1: [f32; MLP_HIDDEN],
-    /// Mask MLP layer 2 weights: hidden (16) → mask_features (32)
-    pub mask_mlp_w2: [[f32; EMBED_DIM]; MLP_HIDDEN],
-    /// Mask MLP layer 2 bias
-    pub mask_mlp_b2: [f32; EMBED_DIM],
-
-    /// Rule MLP layer 1 weights: rule_features (8) → hidden (16).
-    /// Shared across all rules - scales sublinearly with rule count.
-    /// (Legacy: used with hand-crafted RuleFeatures)
-    pub rule_mlp_w1: [[f32; MLP_HIDDEN]; RULE_FEATURE_DIM],
-    /// Rule MLP layer 1 bias
-    pub rule_mlp_b1: [f32; MLP_HIDDEN],
-    /// Rule MLP layer 2 weights: hidden (16) → rule_embed (32)
-    pub rule_mlp_w2: [[f32; EMBED_DIM]; MLP_HIDDEN],
-    /// Rule MLP layer 2 bias
-    pub rule_mlp_b2: [f32; EMBED_DIM],
-
-    // ========== RULE TEMPLATE PROJECTION (LHS/RHS embeddings) ==========
-    // These fields support encoding rules from their LHS/RHS expression
-    // templates using the SAME expr_embed as extraction/saturation heads.
-    //
-    // 4-way concat: [z_LHS | z_RHS | z_LHS-z_RHS | z_LHS*z_RHS] (128) → rule_embed (32)
-    /// Rule projection weights: [RULE_CONCAT_DIM x EMBED_DIM] = [128 x 32] = 4,096 params.
-    /// Projects 4-way concatenation to rule embedding.
-    pub rule_proj_w: [[f32; EMBED_DIM]; RULE_CONCAT_DIM],
-    /// Rule projection bias: [EMBED_DIM] = 32 params
-    pub rule_proj_b: [f32; EMBED_DIM],
-
-    /// Bilinear interaction matrix: mask_features @ interaction @ rule_embed
-    pub interaction: [[f32; EMBED_DIM]; EMBED_DIM],
-
-    /// Learned bias projection: produces per-rule bias via dot(mask_bias_proj, rule_embed)
-    pub mask_bias_proj: [f32; EMBED_DIM],
-
-    // ========== GRAPH STATE BACKBONE (for saturation head) ==========
-    // Separate pathway: GraphAccumulator (VSA e-graph state) → graph_w1 → graph_proj → mask_mlp → bilinear
-    // The extraction head path (EdgeAccumulator → w1 → expr_proj) is completely unchanged.
-    /// Graph backbone weights: [GRAPH_INPUT_DIM][HIDDEN_DIM] (132 × 64 = 8,448 params)
-    pub graph_w1: [[f32; HIDDEN_DIM]; GRAPH_INPUT_DIM],
-    /// Graph backbone biases: [HIDDEN_DIM] (64 params)
-    pub graph_b1: [f32; HIDDEN_DIM],
-    /// Graph → embed projection weights: [HIDDEN_DIM][EMBED_DIM] (64 × 32 = 2,048 params)
-    pub graph_proj_w: [[f32; EMBED_DIM]; HIDDEN_DIM],
-    /// Graph → embed projection bias: [EMBED_DIM] (32 params)
-    pub graph_proj_b: [f32; EMBED_DIM],
 }
 
 impl Default for ExprNnue {
@@ -1753,20 +1232,17 @@ impl Default for ExprNnue {
 }
 
 impl ExprNnue {
-    /// Create a zero-initialized dual-head network.
+    /// Create a zero-initialized extraction head.
     #[must_use]
     pub fn new() -> Self {
         Self {
-            // Backbone
             embeddings: OpEmbeddings::new(),
             w1: [[0.0; HIDDEN_DIM]; INPUT_DIM],
             b1: [0.0; HIDDEN_DIM],
 
-            // Shared trunk (zero-init)
             trunk_w: [[0.0; HIDDEN_DIM]; HIDDEN_DIM],
             trunk_b: [0.0; HIDDEN_DIM],
 
-            // Unified mask architecture
             expr_proj_w: [[0.0; EMBED_DIM]; HIDDEN_DIM],
             expr_proj_b: [0.0; EMBED_DIM],
 
@@ -1774,33 +1250,10 @@ impl ExprNnue {
             value_mlp_b1: [0.0; MLP_HIDDEN],
             value_mlp_w2: [0.0; MLP_HIDDEN],
             value_mlp_b2: 5.0, // Start near typical log-cost
-
-            mask_mlp_w1: [[0.0; MLP_HIDDEN]; MASK_INPUT_DIM], // 24 × 16
-            mask_mlp_b1: [0.0; MLP_HIDDEN],
-            mask_mlp_w2: [[0.0; EMBED_DIM]; MLP_HIDDEN],
-            mask_mlp_b2: [0.0; EMBED_DIM],
-
-            rule_mlp_w1: [[0.0; MLP_HIDDEN]; RULE_FEATURE_DIM],
-            rule_mlp_b1: [0.0; MLP_HIDDEN],
-            rule_mlp_w2: [[0.0; EMBED_DIM]; MLP_HIDDEN],
-            rule_mlp_b2: [0.0; EMBED_DIM],
-
-            // Rule template projection (LHS/RHS embeddings)
-            rule_proj_w: [[0.0; EMBED_DIM]; RULE_CONCAT_DIM],
-            rule_proj_b: [0.0; EMBED_DIM],
-
-            interaction: [[0.0; EMBED_DIM]; EMBED_DIM],
-            mask_bias_proj: [0.0; EMBED_DIM],
-
-            // Graph state backbone (zero-init)
-            graph_w1: [[0.0; HIDDEN_DIM]; GRAPH_INPUT_DIM],
-            graph_b1: [0.0; HIDDEN_DIM],
-            graph_proj_w: [[0.0; EMBED_DIM]; HIDDEN_DIM],
-            graph_proj_b: [0.0; EMBED_DIM],
         }
     }
 
-    /// Create a randomly initialized dual-head network.
+    /// Create a randomly initialized extraction head.
     #[must_use]
     pub fn new_random(seed: u64) -> Self {
         let mut net = Self::new();
@@ -1818,64 +1271,6 @@ impl ExprNnue {
         let mut net = Self::new();
         net.embeddings.init_with_latency_prior(seed);
         net.randomize_weights_only(seed);
-        net
-    }
-
-    /// Convert an older extraction-first model into the unified architecture.
-    ///
-    /// This preserves the edge tower exactly by initializing the shared trunk
-    /// to identity, so pre-trunk hidden activations pass through unchanged.
-    /// Task-specific unified-head weights remain zero-initialized and require
-    /// training.
-    #[must_use]
-    pub fn from_factored(factored: &ExprNnue) -> Self {
-        let mut net = Self {
-            embeddings: factored.embeddings.clone(),
-            w1: factored.w1,
-            b1: factored.b1,
-
-            // Shared trunk starts as identity so the migrated edge tower
-            // preserves its pre-trunk representation exactly.
-            trunk_w: [[0.0; HIDDEN_DIM]; HIDDEN_DIM],
-            trunk_b: [0.0; HIDDEN_DIM],
-
-            // Unified mask architecture - zero-initialized (needs training)
-            expr_proj_w: [[0.0; EMBED_DIM]; HIDDEN_DIM],
-            expr_proj_b: [0.0; EMBED_DIM],
-
-            value_mlp_w1: [[0.0; MLP_HIDDEN]; EMBED_DIM],
-            value_mlp_b1: [0.0; MLP_HIDDEN],
-            value_mlp_w2: [0.0; MLP_HIDDEN],
-            value_mlp_b2: 5.0,
-
-            mask_mlp_w1: [[0.0; MLP_HIDDEN]; MASK_INPUT_DIM], // 24 × 16
-            mask_mlp_b1: [0.0; MLP_HIDDEN],
-            mask_mlp_w2: [[0.0; EMBED_DIM]; MLP_HIDDEN],
-            mask_mlp_b2: [0.0; EMBED_DIM],
-
-            rule_mlp_w1: [[0.0; MLP_HIDDEN]; RULE_FEATURE_DIM],
-            rule_mlp_b1: [0.0; MLP_HIDDEN],
-            rule_mlp_w2: [[0.0; EMBED_DIM]; MLP_HIDDEN],
-            rule_mlp_b2: [0.0; EMBED_DIM],
-
-            // Rule template projection - zero-initialized (needs training)
-            rule_proj_w: [[0.0; EMBED_DIM]; RULE_CONCAT_DIM],
-            rule_proj_b: [0.0; EMBED_DIM],
-
-            interaction: [[0.0; EMBED_DIM]; EMBED_DIM],
-            mask_bias_proj: [0.0; EMBED_DIM],
-
-            // Graph state backbone - zero-initialized (needs training)
-            graph_w1: [[0.0; HIDDEN_DIM]; GRAPH_INPUT_DIM],
-            graph_b1: [0.0; HIDDEN_DIM],
-            graph_proj_w: [[0.0; EMBED_DIM]; HIDDEN_DIM],
-            graph_proj_b: [0.0; EMBED_DIM],
-        };
-
-        for i in 0..HIDDEN_DIM {
-            net.trunk_w[i][i] = 1.0;
-        }
-
         net
     }
 
@@ -1910,20 +1305,18 @@ impl ExprNnue {
             *b = 0.0;
         }
 
-        // Initialize unified mask architecture (full init - includes shared projection + value mlp)
-        self.randomize_unified_arch_with_rng(&mut next_f32);
+        // Initialize the shared projection + value MLP.
+        self.randomize_extraction_head_with_rng(&mut next_f32);
     }
 
-    /// Internal helper to randomize ALL unified architecture weights.
+    /// Internal helper to randomize the expr projection + value MLP.
     ///
-    /// ONLY used during full random init (randomize_weights_only).
-    /// Do NOT call this when bootstrapping from extraction head weights - use randomize_mask_only instead.
-    fn randomize_unified_arch_with_rng<F: FnMut() -> f32>(&mut self, next_f32: &mut F) {
+    /// ONLY used during full random init (`randomize_weights_only`).
+    fn randomize_extraction_head_with_rng<F: FnMut() -> f32>(&mut self, next_f32: &mut F) {
         // He initialization scales
         let scale_proj = sqrtf(2.0 / HIDDEN_DIM as f32);
         let scale_embed = sqrtf(2.0 / EMBED_DIM as f32);
         let scale_hidden = sqrtf(2.0 / MLP_HIDDEN as f32);
-        let scale_rule_feat = sqrtf(2.0 / RULE_FEATURE_DIM as f32);
 
         // Expr projection: HIDDEN_DIM → EMBED_DIM
         for j in 0..HIDDEN_DIM {
@@ -1948,90 +1341,6 @@ impl ExprNnue {
             self.value_mlp_w2[j] = next_f32() * scale_hidden;
         }
         self.value_mlp_b2 = 5.0; // Start near typical log-cost
-
-        // Mask MLP: EMBED_DIM (24) → MLP_HIDDEN → EMBED_DIM
-        let scale_mask_input = sqrtf(2.0 / MASK_INPUT_DIM as f32);
-        for i in 0..MASK_INPUT_DIM {
-            for j in 0..MLP_HIDDEN {
-                self.mask_mlp_w1[i][j] = next_f32() * scale_mask_input;
-            }
-        }
-        for b in &mut self.mask_mlp_b1 {
-            *b = next_f32().abs() * 0.1;
-        }
-        for j in 0..MLP_HIDDEN {
-            for k in 0..EMBED_DIM {
-                self.mask_mlp_w2[j][k] = next_f32() * scale_hidden;
-            }
-        }
-        for b in &mut self.mask_mlp_b2 {
-            *b = 0.0; // Neutral
-        }
-
-        // Rule MLP: RULE_FEATURE_DIM → MLP_HIDDEN → EMBED_DIM (hand-crafted features)
-        for i in 0..RULE_FEATURE_DIM {
-            for j in 0..MLP_HIDDEN {
-                self.rule_mlp_w1[i][j] = next_f32() * scale_rule_feat;
-            }
-        }
-        for b in &mut self.rule_mlp_b1 {
-            *b = next_f32().abs() * 0.1;
-        }
-        for j in 0..MLP_HIDDEN {
-            for k in 0..EMBED_DIM {
-                self.rule_mlp_w2[j][k] = next_f32() * scale_hidden;
-            }
-        }
-        for b in &mut self.rule_mlp_b2 {
-            *b = 0.0; // Neutral
-        }
-
-        // Rule Projection: RULE_CONCAT_DIM (96) → EMBED_DIM (24)
-        // Linear projection from 4-way concat [z_LHS | z_RHS | z_LHS-z_RHS | z_LHS*z_RHS]
-        let scale_concat = sqrtf(2.0 / RULE_CONCAT_DIM as f32);
-        for i in 0..RULE_CONCAT_DIM {
-            for k in 0..EMBED_DIM {
-                self.rule_proj_w[i][k] = next_f32() * scale_concat;
-            }
-        }
-        for b in &mut self.rule_proj_b {
-            *b = 0.0; // Neutral
-        }
-
-        // Interaction matrix: start near identity (simple dot product baseline)
-        for i in 0..EMBED_DIM {
-            for j in 0..EMBED_DIM {
-                self.interaction[i][j] = if i == j { 1.0 } else { next_f32() * 0.1 };
-            }
-        }
-
-        // Bias projection: neutral
-        for b in &mut self.mask_bias_proj {
-            *b = 0.0;
-        }
-
-        // Graph backbone: GRAPH_INPUT_DIM → HIDDEN_DIM
-        let scale_graph = sqrtf(2.0 / GRAPH_INPUT_DIM as f32);
-        for row in 0..GRAPH_INPUT_DIM {
-            for col in 0..HIDDEN_DIM {
-                self.graph_w1[row][col] = next_f32() * scale_graph;
-            }
-        }
-        for b in &mut self.graph_b1 {
-            *b = next_f32().abs() * 0.1;
-        }
-
-        // Graph projection: HIDDEN_DIM → EMBED_DIM
-        for j in 0..HIDDEN_DIM {
-            for k in 0..EMBED_DIM {
-                self.graph_proj_w[j][k] = next_f32() * scale_proj;
-            }
-        }
-        for b in &mut self.graph_proj_b {
-            *b = next_f32().abs() * 0.1;
-        }
-
-        // NOTE: search_mlp randomization removed - mask IS the policy
     }
 
     /// Randomize all weights including embeddings.
@@ -2040,130 +1349,16 @@ impl ExprNnue {
         self.randomize_weights_only(seed);
     }
 
-    /// Create a copy with trained backbone but randomized mask weights.
+    /// Apply the shared trunk: `HIDDEN_DIM -> HIDDEN_DIM` with ReLU.
     ///
-    /// This is the key method for embedding sharing: load a trained extraction head,
-    /// then create a new model that:
-    /// - Keeps: embeddings, w1, b1 (trained backbone)
-    /// - Keeps: expr_proj_w, expr_proj_b (shared projection - trained with extraction head)
-    /// - Keeps: value_mlp_* (extraction head weights)
-    /// - Randomizes: mask_mlp, rule_mlp, rule_proj, interaction, mask_bias_proj (saturation head specific)
-    ///
-    /// Use this when bootstrapping saturation head training from a pre-trained extraction head.
-    #[must_use]
-    pub fn with_randomized_mask_weights(&self, seed: u64) -> Self {
-        let mut model = self.clone();
-        model.randomize_mask_only(seed);
-        model
-    }
-
-    /// Randomize ONLY mask-specific weights, preserving shared backbone and value MLP.
-    ///
-    /// Randomizes: mask_mlp, rule_mlp, rule_proj, interaction, mask_bias_proj
-    /// Preserves: embeddings, w1, b1, expr_proj, value_mlp
-    pub fn randomize_mask_only(&mut self, seed: u64) {
-        let mut rng_state = seed.wrapping_add(54321);
-
-        let mut next_f32 = || {
-            rng_state = rng_state.wrapping_mul(6364136223846793005).wrapping_add(1);
-            (rng_state >> 33) as f32 / (1u64 << 31) as f32 * 2.0 - 1.0
-        };
-
-        // He initialization scales
-        let scale_embed = sqrtf(2.0 / EMBED_DIM as f32);
-        let scale_mask_input = sqrtf(2.0 / MASK_INPUT_DIM as f32); // 24 dims
-        let scale_hidden = sqrtf(2.0 / MLP_HIDDEN as f32);
-        let scale_rule_feat = sqrtf(2.0 / RULE_FEATURE_DIM as f32);
-        let scale_concat = sqrtf(2.0 / RULE_CONCAT_DIM as f32);
-
-        // Mask MLP: EMBED_DIM (24) → MLP_HIDDEN → EMBED_DIM
-        for i in 0..MASK_INPUT_DIM {
-            for j in 0..MLP_HIDDEN {
-                self.mask_mlp_w1[i][j] = next_f32() * scale_mask_input;
-            }
-        }
-        for b in &mut self.mask_mlp_b1 {
-            *b = next_f32().abs() * 0.1;
-        }
-        for j in 0..MLP_HIDDEN {
-            for k in 0..EMBED_DIM {
-                self.mask_mlp_w2[j][k] = next_f32() * scale_hidden;
-            }
-        }
-        for b in &mut self.mask_mlp_b2 {
-            *b = 0.0;
-        }
-
-        // Rule MLP: RULE_FEATURE_DIM → MLP_HIDDEN → EMBED_DIM (hand-crafted features)
-        for i in 0..RULE_FEATURE_DIM {
-            for j in 0..MLP_HIDDEN {
-                self.rule_mlp_w1[i][j] = next_f32() * scale_rule_feat;
-            }
-        }
-        for b in &mut self.rule_mlp_b1 {
-            *b = next_f32().abs() * 0.1;
-        }
-        for j in 0..MLP_HIDDEN {
-            for k in 0..EMBED_DIM {
-                self.rule_mlp_w2[j][k] = next_f32() * scale_hidden;
-            }
-        }
-        for b in &mut self.rule_mlp_b2 {
-            *b = 0.0;
-        }
-
-        // Rule Projection: RULE_CONCAT_DIM (96) → EMBED_DIM (24)
-        for i in 0..RULE_CONCAT_DIM {
-            for k in 0..EMBED_DIM {
-                self.rule_proj_w[i][k] = next_f32() * scale_concat;
-            }
-        }
-        for b in &mut self.rule_proj_b {
-            *b = 0.0;
-        }
-
-        // Interaction matrix: start near identity
-        for i in 0..EMBED_DIM {
-            for j in 0..EMBED_DIM {
-                self.interaction[i][j] = if i == j { 1.0 } else { next_f32() * 0.1 };
-            }
-        }
-
-        // Bias projection: neutral
-        for b in &mut self.mask_bias_proj {
-            *b = 0.0;
-        }
-
-        // Graph backbone: GRAPH_INPUT_DIM → HIDDEN_DIM (mask-specific pathway)
-        let scale_graph = sqrtf(2.0 / GRAPH_INPUT_DIM as f32);
-        let scale_proj = sqrtf(2.0 / HIDDEN_DIM as f32);
-        for row in 0..GRAPH_INPUT_DIM {
-            for col in 0..HIDDEN_DIM {
-                self.graph_w1[row][col] = next_f32() * scale_graph;
-            }
-        }
-        for b in &mut self.graph_b1 {
-            *b = next_f32().abs() * 0.1;
-        }
-
-        // Graph projection: HIDDEN_DIM → EMBED_DIM
-        for j in 0..HIDDEN_DIM {
-            for k in 0..EMBED_DIM {
-                self.graph_proj_w[j][k] = next_f32() * scale_proj;
-            }
-        }
-        for b in &mut self.graph_proj_b {
-            *b = next_f32().abs() * 0.1;
-        }
-    }
-
-    /// Apply shared trunk: HIDDEN_DIM -> HIDDEN_DIM with ReLU.
-    ///
-    /// Both the edge tower and graph tower pass through this layer before
-    /// reaching their task-specific projection heads. Initialized near-identity
-    /// so the trunk preserves tower signal until training pulls it away.
+    /// Both this struct's edge tower and `nnue::guide`'s graph tower pass
+    /// through this layer before reaching their task-specific projection
+    /// heads — `pub(crate)` rather than private so the guide module (same
+    /// crate, different module) can share it instead of duplicating it.
+    /// Initialized near-identity so the trunk preserves tower signal until
+    /// training pulls it away.
     #[inline]
-    fn apply_trunk(&self, tower_output: &[f32; HIDDEN_DIM]) -> [f32; HIDDEN_DIM] {
+    pub(crate) fn apply_trunk(&self, tower_output: &[f32; HIDDEN_DIM]) -> [f32; HIDDEN_DIM] {
         let mut out = self.trunk_b;
         for i in 0..HIDDEN_DIM {
             for j in 0..HIDDEN_DIM {
@@ -2221,30 +1416,6 @@ impl ExprNnue {
         hidden
     }
 
-    // ========================================================================
-    // Unified Mask Architecture Forward Pass
-    //
-    // New bilinear interaction model that scales to 1000+ rules.
-    //
-    // Architecture:
-    //   expr → backbone → hidden → expr_proj → expr_embed (shared)
-    //                                              │
-    //          ┌───────────────────────────────────┼───────────────────┐
-    //          │                                   │                   │
-    //          ▼                                   ▼                   ▼
-    //    value_mlp (private)               mask_mlp (private)    rule_features
-    //          │                                   │                   │
-    //          ▼                                   │             rule_mlp (shared)
-    //       cost (1)                               │                   │
-    //                                              ▼                   ▼
-    //                                        mask_features       rule_embed
-    //                                              │                   │
-    //                                              └──── bilinear ─────┘
-    //                                                       │
-    //                                                       ▼
-    //                                              score + rule_bias
-    // ========================================================================
-
     /// Project backbone hidden to shared expr embedding (EMBED_DIM).
     #[inline]
     pub fn compute_expr_embed(&self, hidden: &[f32; HIDDEN_DIM]) -> [f32; EMBED_DIM] {
@@ -2255,129 +1426,6 @@ impl ExprNnue {
             }
         }
         embed
-    }
-
-    // ========================================================================
-    // Graph State Backbone Forward Pass
-    //
-    // Separate pathway for saturation head: GraphAccumulator → graph_w1 → graph_proj
-    // Feeds into the SAME mask_mlp + bilinear scoring as the expr pathway.
-    // ========================================================================
-
-    /// Graph state forward pass (for saturation head).
-    ///
-    /// Same structure as `forward_shared` but with `graph_w1`/`graph_b1` and
-    /// `GRAPH_INPUT_DIM` input (132). Uses `1/sqrt(node_count)` scaling
-    /// and log2 scalars, matching the `forward_shared` conventions.
-    #[inline]
-    pub fn forward_graph(&self, gacc: &GraphAccumulator) -> [f32; HIDDEN_DIM] {
-        let mut hidden = self.graph_b1;
-
-        // Scale factor: 1/sqrt(N) prevents variance explosion from summing N embeddings.
-        let scale = if gacc.node_count > 0 {
-            1.0 / libm::sqrtf(gacc.node_count as f32)
-        } else {
-            1.0
-        };
-
-        // Process graph accumulator (128 dims: 4K sections)
-        for (i, &val) in gacc.values.iter().enumerate() {
-            let scaled_val = val * scale;
-            for (j, h) in hidden.iter_mut().enumerate() {
-                *h += scaled_val * self.graph_w1[i][j];
-            }
-        }
-
-        // Process scalar features (4 dims: edge_count, node_count, node_budget, epoch_budget).
-        // Use log2 to compress the range for large e-graphs.
-        let base = GRAPH_ACC_DIM;
-        let ec = libm::log2f(1.0 + gacc.edge_count as f32);
-        let nc = libm::log2f(1.0 + gacc.node_count as f32);
-        let nb = libm::log2f(1.0 + gacc.node_budget as f32);
-        let eb = libm::log2f(1.0 + gacc.epoch_budget as f32);
-        for (j, h) in hidden.iter_mut().enumerate() {
-            *h += ec * self.graph_w1[base][j];
-            *h += nc * self.graph_w1[base + 1][j];
-            *h += nb * self.graph_w1[base + 2][j];
-            *h += eb * self.graph_w1[base + 3][j];
-        }
-
-        // ReLU activation
-        for h in &mut hidden {
-            *h = h.max(0.0);
-        }
-
-        // Shared trunk
-        let hidden = self.apply_trunk(&hidden);
-
-        hidden
-    }
-
-    /// Project graph hidden to graph embedding (EMBED_DIM).
-    ///
-    /// Same structure as `compute_expr_embed` but with `graph_proj_w`/`graph_proj_b`.
-    #[inline]
-    pub fn compute_graph_embed(&self, hidden: &[f32; HIDDEN_DIM]) -> [f32; EMBED_DIM] {
-        let mut embed = self.graph_proj_b;
-        for j in 0..HIDDEN_DIM {
-            for k in 0..EMBED_DIM {
-                embed[k] += hidden[j] * self.graph_proj_w[j][k];
-            }
-        }
-        embed
-    }
-
-    /// Score all rules using graph state (not expression state).
-    ///
-    /// `forward_graph → compute_graph_embed → compute_mask_features → bilinear_score`
-    ///
-    /// The mask_mlp, interaction matrix, and bilinear_score are **shared** with the
-    /// expression pathway — only the input pathway changes.
-    #[must_use]
-    pub fn mask_score_all_rules_graph(
-        &self,
-        gacc: &GraphAccumulator,
-        rule_embeds: &[[f32; EMBED_DIM]],
-    ) -> Vec<f32> {
-        let hidden = self.forward_graph(gacc);
-        let graph_embed = self.compute_graph_embed(&hidden);
-        let mask_features = self.compute_mask_features(&graph_embed);
-        rule_embeds
-            .iter()
-            .map(|re| self.bilinear_score(&mask_features, re))
-            .collect()
-    }
-
-    /// Compute mask features from expr embedding and value prediction (for bilinear scoring).
-    ///
-    /// Transform expr_embed through mask MLP to produce mask features for bilinear scoring.
-    ///
-    /// Input: expr_embed (24 dims) directly — value_pred was removed as redundant.
-    /// MLP: EMBED_DIM (24) → MLP_HIDDEN (ReLU) → EMBED_DIM (24)
-    #[inline]
-    fn compute_mask_features(&self, expr_embed: &[f32; EMBED_DIM]) -> [f32; EMBED_DIM] {
-        // First layer: EMBED_DIM → MLP_HIDDEN
-        let mut h = self.mask_mlp_b1;
-
-        for i in 0..EMBED_DIM {
-            for j in 0..MLP_HIDDEN {
-                h[j] += expr_embed[i] * self.mask_mlp_w1[i][j];
-            }
-        }
-
-        // ReLU
-        for j in 0..MLP_HIDDEN {
-            h[j] = h[j].max(0.0);
-        }
-
-        // Second layer: MLP_HIDDEN → EMBED_DIM
-        let mut out = self.mask_mlp_b2;
-        for j in 0..MLP_HIDDEN {
-            for k in 0..EMBED_DIM {
-                out[k] += h[j] * self.mask_mlp_w2[j][k];
-            }
-        }
-        out
     }
 
     /// Forward pass through value MLP from expr embedding.
@@ -2407,165 +1455,6 @@ impl ExprNnue {
         cost
     }
 
-    /// Encode rule features to rule embedding.
-    ///
-    /// MLP: RULE_FEATURE_DIM → MLP_HIDDEN (ReLU) → EMBED_DIM
-    /// This MLP is shared across all rules - scales sublinearly with rule count.
-    #[must_use]
-    pub fn encode_rule(&self, rule_features: &[f32; RULE_FEATURE_DIM]) -> [f32; EMBED_DIM] {
-        // First layer: RULE_FEATURE_DIM → MLP_HIDDEN
-        let mut h = self.rule_mlp_b1;
-        for i in 0..RULE_FEATURE_DIM {
-            for j in 0..MLP_HIDDEN {
-                h[j] += rule_features[i] * self.rule_mlp_w1[i][j];
-            }
-        }
-
-        // ReLU
-        for j in 0..MLP_HIDDEN {
-            h[j] = h[j].max(0.0);
-        }
-
-        // Second layer: MLP_HIDDEN → EMBED_DIM
-        let mut out = self.rule_mlp_b2;
-        for j in 0..MLP_HIDDEN {
-            for k in 0..EMBED_DIM {
-                out[k] += h[j] * self.rule_mlp_w2[j][k];
-            }
-        }
-        out
-    }
-
-    /// Pre-encode all rules (call once, cache results).
-    ///
-    /// Returns a Vec of rule embeddings that can be reused across multiple
-    /// expressions during saturation.
-    #[must_use]
-    pub fn encode_all_rules(
-        &self,
-        rule_features: &RuleFeatures,
-        num_rules: usize,
-    ) -> Vec<[f32; EMBED_DIM]> {
-        (0..num_rules)
-            .map(|r| self.encode_rule(&rule_features.features[r]))
-            .collect()
-    }
-
-    // =========================================================================
-    // Rule Encoding from LHS/RHS Templates
-    //
-    // Uses the SAME expr_embed as extraction/saturation heads. 4-way concatenation:
-    // [z_LHS | z_RHS | z_LHS-z_RHS | z_LHS*z_RHS] → linear → rule_embed
-    //
-    // This provides richer semantic features than hand-crafted rule descriptors.
-    // =========================================================================
-
-    /// Pre-encode all rules from templates (call once at init, cache results).
-    ///
-    /// Rules without templates fall back to zero embedding.
-    /// Rule embeddings don't change during search - they're computed from
-    /// LHS/RHS templates which are static.
-    #[must_use]
-    pub fn encode_all_rules_from_templates(
-        &self,
-        templates: &RuleTemplates,
-    ) -> Vec<[f32; EMBED_DIM]> {
-        (0..templates.len())
-            .map(|r| match templates.get(r) {
-                Some(t) => match (t.lhs, t.rhs) {
-                    (Some(lhs), Some(rhs)) => self.encode_rule_from_arena(&t.arena, lhs, rhs),
-                    _ => [0.0f32; EMBED_DIM],
-                },
-                None => [0.0f32; EMBED_DIM], // No template - zero embedding
-            })
-            .collect()
-    }
-
-    /// Encode a single rule's LHS/RHS arena subtrees into a rule embedding.
-    ///
-    /// 4-way concatenation `[z_LHS | z_RHS | z_LHS-z_RHS | z_LHS*z_RHS]` projected
-    /// to `EMBED_DIM`, using the same shared backbone as the extraction head.
-    #[must_use]
-    pub fn encode_rule_from_arena(
-        &self,
-        arena: &ExprArena,
-        lhs: ExprId,
-        rhs: ExprId,
-    ) -> [f32; EMBED_DIM] {
-        // Same feature path as everything else that touches `w1`: the slot
-        // semantics of `extraction_input` (variance fractions in 4K..4K+4)
-        // must have exactly one meaning across the whole net.
-        let lhs_acc = EdgeAccumulator::from_arena_dag(arena, lhs, &self.embeddings);
-        let lhs_hidden = self.forward_expr_only(&lhs_acc);
-        let z_lhs = self.compute_expr_embed(&lhs_hidden);
-
-        let rhs_acc = EdgeAccumulator::from_arena_dag(arena, rhs, &self.embeddings);
-        let rhs_hidden = self.forward_expr_only(&rhs_acc);
-        let z_rhs = self.compute_expr_embed(&rhs_hidden);
-
-        let mut concat = [0.0f32; RULE_CONCAT_DIM];
-        for i in 0..EMBED_DIM {
-            concat[i] = z_lhs[i];
-            concat[EMBED_DIM + i] = z_rhs[i];
-            concat[2 * EMBED_DIM + i] = z_lhs[i] - z_rhs[i];
-            concat[3 * EMBED_DIM + i] = z_lhs[i] * z_rhs[i];
-        }
-
-        let mut out = self.rule_proj_b;
-        for i in 0..RULE_CONCAT_DIM {
-            for k in 0..EMBED_DIM {
-                out[k] += concat[i] * self.rule_proj_w[i][k];
-            }
-        }
-        out
-    }
-
-    /// Bilinear score: mask_features @ interaction @ rule_embed + bias.
-    ///
-    /// Efficient O(1) scoring with pre-computed mask_features.
-    #[inline]
-    #[must_use]
-    pub fn bilinear_score(
-        &self,
-        mask_features: &[f32; EMBED_DIM],
-        rule_embed: &[f32; EMBED_DIM],
-    ) -> f32 {
-        // transformed = mask_features @ interaction
-        let mut transformed = [0.0f32; EMBED_DIM];
-        for i in 0..EMBED_DIM {
-            for j in 0..EMBED_DIM {
-                transformed[j] += mask_features[i] * self.interaction[i][j];
-            }
-        }
-
-        // score = dot(transformed + mask_bias_proj, rule_embed)
-        let mut score = 0.0f32;
-        for k in 0..EMBED_DIM {
-            score += (transformed[k] + self.mask_bias_proj[k]) * rule_embed[k];
-        }
-        score
-    }
-
-    /// Score all rules with pre-computed backbone hidden state.
-    ///
-    /// More efficient when you have multiple expressions sharing the same
-    /// feature extraction.
-    #[must_use]
-    pub fn mask_score_all_rules_with_hidden(
-        &self,
-        hidden: &[f32; HIDDEN_DIM],
-        rule_embeds: &[[f32; EMBED_DIM]],
-    ) -> Vec<f32> {
-        let expr_embed = self.compute_expr_embed(hidden);
-
-        let mask_features = self.compute_mask_features(&expr_embed);
-
-        rule_embeds
-            .iter()
-            .map(|rule_embed| self.bilinear_score(&mask_features, rule_embed))
-            .collect()
-    }
-
     /// Total parameter count.
     #[must_use]
     pub const fn param_count() -> usize {
@@ -2582,28 +1471,7 @@ impl ExprNnue {
             + EMBED_DIM * MLP_HIDDEN          // value_mlp_w1: 32 * 16 = 512
             + MLP_HIDDEN                      // value_mlp_b1: 16
             + MLP_HIDDEN                      // value_mlp_w2: 16
-            + 1                               // value_mlp_b2: 1
-            // mask MLP
-            + MASK_INPUT_DIM * MLP_HIDDEN     // mask_mlp_w1: 32 * 16 = 512
-            + MLP_HIDDEN                      // mask_mlp_b1: 16
-            + MLP_HIDDEN * EMBED_DIM          // mask_mlp_w2: 16 * 32 = 512
-            + EMBED_DIM                       // mask_mlp_b2: 32
-            // rule MLP
-            + RULE_FEATURE_DIM * MLP_HIDDEN   // rule_mlp_w1: 8 * 16 = 128
-            + MLP_HIDDEN                      // rule_mlp_b1: 16
-            + MLP_HIDDEN * EMBED_DIM          // rule_mlp_w2: 16 * 32 = 512
-            + EMBED_DIM                       // rule_mlp_b2: 32
-            // rule projection
-            + RULE_CONCAT_DIM * EMBED_DIM     // rule_proj_w: 128 * 32 = 4,096
-            + EMBED_DIM                       // rule_proj_b: 32
-            // bilinear
-            + EMBED_DIM * EMBED_DIM           // interaction: 32 * 32 = 1,024
-            + EMBED_DIM                        // mask_bias_proj: 32
-            // graph state backbone
-            + GRAPH_INPUT_DIM * HIDDEN_DIM    // graph_w1: 132 * 64 = 8,448
-            + HIDDEN_DIM                      // graph_b1: 64
-            + HIDDEN_DIM * EMBED_DIM          // graph_proj_w: 64 * 32 = 2,048
-            + EMBED_DIM // graph_proj_b: 32
+            + 1 // value_mlp_b2: 1
     }
 
     /// Memory size in bytes (f32 weights).
@@ -2618,13 +1486,21 @@ impl ExprNnue {
     // in the workspace, and they were the last consumers of the old
     // `forward_shared` path that fed log2 count/budget scalars into the same
     // `w1` slots the extraction head uses for the variance histogram — the
-    // exact train/deploy skew this round removed. The saturation head scores
-    // rules through the graph tower (`forward_graph`), not the edge tower.
+    // exact train/deploy skew this round removed. The saturation head (now
+    // `nnue::guide`, its own weights entirely) scores rules through its own
+    // graph tower, not this struct's edge tower.
 
     /// Save weights to a binary file.
     ///
-    /// Format: magic "TRIE" + all weights as little-endian f32.
-    /// TRIE: op embeddings are one row per op, in `OpKind`'s order.
+    /// Format: magic "TRIF" + all weights as little-endian f32.
+    /// TRIF: **live-only** — this checkpoint holds exactly the params
+    /// [`Self::param_count`] counts (the extraction head + shared backbone,
+    /// 16,897 params / 67,588 bytes). Earlier formats (TRIE and before) also
+    /// serialized the saturation head's mask/rule/graph weights — untrained
+    /// noise that made up roughly half the file by byte count and drifted
+    /// toward zero every training run under weight decay with no gradient
+    /// ever opposing it, since nothing trained them. Op embeddings are one
+    /// row per op, in `OpKind`'s order.
     #[cfg(feature = "std")]
     pub fn save(&self, path: &std::path::Path) -> std::io::Result<()> {
         use std::io::Write;
@@ -2632,7 +1508,7 @@ impl ExprNnue {
 
         // Magic header. Bump it whenever these bytes change meaning — including
         // when `OpKind`'s order changes, since the embedding rows follow it.
-        file.write_all(b"TRIE")?;
+        file.write_all(b"TRIF")?;
 
         // ===== Backbone =====
         // Embeddings
@@ -2662,7 +1538,6 @@ impl ExprNnue {
             file.write_all(&val.to_le_bytes())?;
         }
 
-        // ===== Unified mask architecture =====
         // Expr projection
         for row in &self.expr_proj_w {
             for &val in row {
@@ -2687,82 +1562,6 @@ impl ExprNnue {
         }
         file.write_all(&self.value_mlp_b2.to_le_bytes())?;
 
-        // Mask MLP
-        for row in &self.mask_mlp_w1 {
-            for &val in row {
-                file.write_all(&val.to_le_bytes())?;
-            }
-        }
-        for &val in &self.mask_mlp_b1 {
-            file.write_all(&val.to_le_bytes())?;
-        }
-        for row in &self.mask_mlp_w2 {
-            for &val in row {
-                file.write_all(&val.to_le_bytes())?;
-            }
-        }
-        for &val in &self.mask_mlp_b2 {
-            file.write_all(&val.to_le_bytes())?;
-        }
-
-        // Rule MLP for hand-crafted rule features.
-        for row in &self.rule_mlp_w1 {
-            for &val in row {
-                file.write_all(&val.to_le_bytes())?;
-            }
-        }
-        for &val in &self.rule_mlp_b1 {
-            file.write_all(&val.to_le_bytes())?;
-        }
-        for row in &self.rule_mlp_w2 {
-            for &val in row {
-                file.write_all(&val.to_le_bytes())?;
-            }
-        }
-        for &val in &self.rule_mlp_b2 {
-            file.write_all(&val.to_le_bytes())?;
-        }
-
-        // Rule Projection (TRI3: LHS/RHS template encoding)
-        for row in &self.rule_proj_w {
-            for &val in row {
-                file.write_all(&val.to_le_bytes())?;
-            }
-        }
-        for &val in &self.rule_proj_b {
-            file.write_all(&val.to_le_bytes())?;
-        }
-
-        // Interaction matrix
-        for row in &self.interaction {
-            for &val in row {
-                file.write_all(&val.to_le_bytes())?;
-            }
-        }
-
-        // Mask bias projection
-        for &val in &self.mask_bias_proj {
-            file.write_all(&val.to_le_bytes())?;
-        }
-
-        // Graph state backbone
-        for row in &self.graph_w1 {
-            for &val in row {
-                file.write_all(&val.to_le_bytes())?;
-            }
-        }
-        for &val in &self.graph_b1 {
-            file.write_all(&val.to_le_bytes())?;
-        }
-        for row in &self.graph_proj_w {
-            for &val in row {
-                file.write_all(&val.to_le_bytes())?;
-            }
-        }
-        for &val in &self.graph_proj_b {
-            file.write_all(&val.to_le_bytes())?;
-        }
-
         Ok(())
     }
 
@@ -2776,7 +1575,7 @@ impl ExprNnue {
 
     /// Load weights from a binary file.
     ///
-    /// Only supports "TRIE" format. Older formats (TRID and earlier) require retrain.
+    /// Only supports "TRIF" format. Older formats (TRIE and earlier) require retrain.
     #[cfg(feature = "std")]
     pub fn load(path: &std::path::Path) -> std::io::Result<Self> {
         let file = std::io::BufReader::with_capacity(256 * 1024, std::fs::File::open(path)?);
@@ -2789,10 +1588,19 @@ impl ExprNnue {
         let mut magic = [0u8; 4];
         file.read_exact(&mut magic)?;
 
-        // TRIE: op embeddings are one row per op, in `OpKind`'s order. That
+        // TRIF: live-only — extraction head + shared backbone ONLY. Op
+        //       embeddings are one row per op, in `OpKind`'s order. That
         //       order is `pixelflow-ir`'s to change, and changing it silently
         //       re-attaches every row to a different operation — so a change
         //       there means a new magic here.
+        // TRIE: incompatible — also serialized the saturation head's
+        //       mask/rule/graph weights (untrained noise, since nothing
+        //       trained them — see the 2026-08-17 cost-model domain-model
+        //       reorganization). Those params now live behind
+        //       `nnue::guide::SaturationGuide`, with their own (unused until
+        //       Phase 3) checkpoint format — retrain the extraction head from
+        //       this file's raw bytes is not possible; a TRIE file is simply
+        //       a different, incompatible layout.
         // TRID: incompatible — op embeddings are indexed by `OpKind::index()`,
         //       and the pre-2026-08-02 discriminants had gaps at 17/31/39. The
         //       shapes still match, so a TRID file would load without error and
@@ -2801,17 +1609,21 @@ impl ExprNnue {
         // TRIB: incompatible — GRAPH_ACC_DIM was 3K (96), GRAPH_INPUT_DIM was 100
         // TRIA: incompatible — had mask_rule_bias[1024] instead of mask_bias_proj[32]
         // TRI5-TRI9: incompatible — EMBED_DIM was 24, all weight shapes differ
-        if &magic != b"TRIE" {
+        if &magic != b"TRIF" {
             return Err(std::io::Error::new(
                 std::io::ErrorKind::InvalidData,
                 format!(
-                    "Incompatible ExprNnue format {:?}. Expected 'TRIE'. {}",
+                    "Incompatible ExprNnue format {:?}. Expected 'TRIF'. {}",
                     std::str::from_utf8(&magic).unwrap_or("????"),
-                    if &magic == b"TRID" {
+                    if &magic == b"TRIE" {
+                        "'TRIE' predates the live-only checkpoint split (2026-08-17): it also \
+                         carries the (untrained) saturation-head weights this format dropped. \
+                         Retrain required."
+                    } else if &magic == b"TRID" {
                         "'TRID' predates the dense op numbering; its embeddings are \
                          attributed to the wrong operations. Retrain required."
                     } else {
-                        "Old formats (TRIC, TRIB, TRIA, TRI5-TRI9) require retrain."
+                        "Old formats (TRIE, TRIC, TRIB, TRIA, TRI5-TRI9) require retrain."
                     }
                 ),
             ));
@@ -2857,7 +1669,6 @@ impl ExprNnue {
             *val = f32::from_le_bytes(buf);
         }
 
-        // ===== Unified mask architecture =====
         // Expr projection
         for row in &mut net.expr_proj_w {
             for val in row {
@@ -2896,467 +1707,7 @@ impl ExprNnue {
             net.value_mlp_b2 = f32::from_le_bytes(buf);
         }
 
-        // Mask MLP
-        for row in &mut net.mask_mlp_w1 {
-            for val in row {
-                let mut buf = [0u8; 4];
-                file.read_exact(&mut buf)?;
-                *val = f32::from_le_bytes(buf);
-            }
-        }
-        for val in &mut net.mask_mlp_b1 {
-            let mut buf = [0u8; 4];
-            file.read_exact(&mut buf)?;
-            *val = f32::from_le_bytes(buf);
-        }
-        for row in &mut net.mask_mlp_w2 {
-            for val in row {
-                let mut buf = [0u8; 4];
-                file.read_exact(&mut buf)?;
-                *val = f32::from_le_bytes(buf);
-            }
-        }
-        for val in &mut net.mask_mlp_b2 {
-            let mut buf = [0u8; 4];
-            file.read_exact(&mut buf)?;
-            *val = f32::from_le_bytes(buf);
-        }
-
-        // Rule MLP
-        for row in &mut net.rule_mlp_w1 {
-            for val in row {
-                let mut buf = [0u8; 4];
-                file.read_exact(&mut buf)?;
-                *val = f32::from_le_bytes(buf);
-            }
-        }
-        for val in &mut net.rule_mlp_b1 {
-            let mut buf = [0u8; 4];
-            file.read_exact(&mut buf)?;
-            *val = f32::from_le_bytes(buf);
-        }
-        for row in &mut net.rule_mlp_w2 {
-            for val in row {
-                let mut buf = [0u8; 4];
-                file.read_exact(&mut buf)?;
-                *val = f32::from_le_bytes(buf);
-            }
-        }
-        for val in &mut net.rule_mlp_b2 {
-            let mut buf = [0u8; 4];
-            file.read_exact(&mut buf)?;
-            *val = f32::from_le_bytes(buf);
-        }
-
-        // Rule Projection (LHS/RHS template encoding)
-        for row in &mut net.rule_proj_w {
-            for val in row {
-                let mut buf = [0u8; 4];
-                file.read_exact(&mut buf)?;
-                *val = f32::from_le_bytes(buf);
-            }
-        }
-        for val in &mut net.rule_proj_b {
-            let mut buf = [0u8; 4];
-            file.read_exact(&mut buf)?;
-            *val = f32::from_le_bytes(buf);
-        }
-
-        // Interaction matrix
-        for row in &mut net.interaction {
-            for val in row {
-                let mut buf = [0u8; 4];
-                file.read_exact(&mut buf)?;
-                *val = f32::from_le_bytes(buf);
-            }
-        }
-
-        // Mask bias projection
-        for val in &mut net.mask_bias_proj {
-            let mut buf = [0u8; 4];
-            file.read_exact(&mut buf)?;
-            *val = f32::from_le_bytes(buf);
-        }
-
-        // Graph state backbone (TRIE format: mandatory, no backward compat)
-        for row in &mut net.graph_w1 {
-            for val in row {
-                let mut buf = [0u8; 4];
-                file.read_exact(&mut buf)?;
-                *val = f32::from_le_bytes(buf);
-            }
-        }
-        for val in &mut net.graph_b1 {
-            let mut buf = [0u8; 4];
-            file.read_exact(&mut buf)?;
-            *val = f32::from_le_bytes(buf);
-        }
-        for row in &mut net.graph_proj_w {
-            for val in row {
-                let mut buf = [0u8; 4];
-                file.read_exact(&mut buf)?;
-                *val = f32::from_le_bytes(buf);
-            }
-        }
-        for val in &mut net.graph_proj_b {
-            let mut buf = [0u8; 4];
-            file.read_exact(&mut buf)?;
-            *val = f32::from_le_bytes(buf);
-        }
-
         Ok(net)
-    }
-
-    // =========================================================================
-    // Training Methods
-    // =========================================================================
-
-    // =========================================================================
-    // Unified Mask Architecture Training
-    //
-    // Backprop path: score → bilinear → mask_mlp → expr_embed → expr_proj
-    // Also updates: rule_mlp, interaction, rule_bias
-    // Backbone (embeddings, w1, b1) is FROZEN during mask training.
-    // =========================================================================
-
-    // =========================================================================
-    // Forward with Hidden (for backprop)
-    // =========================================================================
-
-    /// Mask MLP forward storing hidden activations.
-    fn mask_mlp_forward_with_hidden(
-        &self,
-        expr_embed: &[f32; EMBED_DIM],
-    ) -> ([f32; EMBED_DIM], [f32; MLP_HIDDEN]) {
-        // First layer
-        let mut h = self.mask_mlp_b1;
-        for i in 0..EMBED_DIM {
-            for j in 0..MLP_HIDDEN {
-                h[j] += expr_embed[i] * self.mask_mlp_w1[i][j];
-            }
-        }
-
-        // Store pre-ReLU for backprop (we need to know which neurons were active)
-        let h_pre_relu = h;
-
-        // ReLU
-        for j in 0..MLP_HIDDEN {
-            h[j] = h[j].max(0.0);
-        }
-
-        // Second layer
-        let mut out = self.mask_mlp_b2;
-        for j in 0..MLP_HIDDEN {
-            for k in 0..EMBED_DIM {
-                out[k] += h[j] * self.mask_mlp_w2[j][k];
-            }
-        }
-
-        (out, h_pre_relu)
-    }
-
-    /// Rule MLP forward storing hidden activations.
-    fn rule_mlp_forward_with_hidden(
-        &self,
-        rule_features: &[f32; RULE_FEATURE_DIM],
-    ) -> ([f32; EMBED_DIM], [f32; MLP_HIDDEN]) {
-        // First layer
-        let mut h = self.rule_mlp_b1;
-        for i in 0..RULE_FEATURE_DIM {
-            for j in 0..MLP_HIDDEN {
-                h[j] += rule_features[i] * self.rule_mlp_w1[i][j];
-            }
-        }
-
-        let h_pre_relu = h;
-
-        // ReLU
-        for j in 0..MLP_HIDDEN {
-            h[j] = h[j].max(0.0);
-        }
-
-        // Second layer
-        let mut out = self.rule_mlp_b2;
-        for j in 0..MLP_HIDDEN {
-            for k in 0..EMBED_DIM {
-                out[k] += h[j] * self.rule_mlp_w2[j][k];
-            }
-        }
-
-        (out, h_pre_relu)
-    }
-
-    /// Value MLP forward storing hidden activations.
-    fn value_mlp_forward_with_hidden(
-        &self,
-        expr_embed: &[f32; EMBED_DIM],
-    ) -> (f32, [f32; MLP_HIDDEN]) {
-        // First layer
-        let mut h = self.value_mlp_b1;
-        for i in 0..EMBED_DIM {
-            for j in 0..MLP_HIDDEN {
-                h[j] += expr_embed[i] * self.value_mlp_w1[i][j];
-            }
-        }
-
-        let h_pre_relu = h;
-
-        // ReLU
-        for j in 0..MLP_HIDDEN {
-            h[j] = h[j].max(0.0);
-        }
-
-        // Second layer
-        let mut cost = self.value_mlp_b2;
-        for j in 0..MLP_HIDDEN {
-            cost += h[j] * self.value_mlp_w2[j];
-        }
-
-        (cost, h_pre_relu)
-    }
-
-    /// Bilinear forward storing transformed vector.
-    fn bilinear_forward_with_hidden(
-        &self,
-        mask_features: &[f32; EMBED_DIM],
-        rule_embed: &[f32; EMBED_DIM],
-    ) -> (f32, [f32; EMBED_DIM]) {
-        // transformed = mask_features @ interaction
-        let mut transformed = [0.0f32; EMBED_DIM];
-        for i in 0..EMBED_DIM {
-            for j in 0..EMBED_DIM {
-                transformed[j] += mask_features[i] * self.interaction[i][j];
-            }
-        }
-
-        // score = dot(transformed + mask_bias_proj, rule_embed)
-        let mut score = 0.0f32;
-        for k in 0..EMBED_DIM {
-            score += (transformed[k] + self.mask_bias_proj[k]) * rule_embed[k];
-        }
-
-        (score, transformed)
-    }
-
-    // =========================================================================
-    // Backpropagation Helpers
-    // =========================================================================
-
-    /// Backprop through bilinear layer.
-    ///
-    /// Returns (d_mask_features, d_rule_embed) and updates interaction matrix.
-    fn backprop_bilinear(
-        &mut self,
-        d_score: f32,
-        mask_features: &[f32; EMBED_DIM],
-        rule_embed: &[f32; EMBED_DIM],
-        transformed: &[f32; EMBED_DIM],
-        lr: f32,
-    ) -> ([f32; EMBED_DIM], [f32; EMBED_DIM]) {
-        // d_score/d_transformed = rule_embed
-        // d_score/d_rule_embed = transformed
-        let mut d_transformed = [0.0f32; EMBED_DIM];
-        let mut d_rule_embed = [0.0f32; EMBED_DIM];
-
-        for k in 0..EMBED_DIM {
-            d_transformed[k] = d_score * rule_embed[k];
-            d_rule_embed[k] = d_score * transformed[k];
-        }
-
-        // d_transformed/d_mask_features = interaction^T
-        // d_transformed/d_interaction = outer(mask_features, I)
-        let mut d_mask_features = [0.0f32; EMBED_DIM];
-        for i in 0..EMBED_DIM {
-            for j in 0..EMBED_DIM {
-                d_mask_features[i] += d_transformed[j] * self.interaction[i][j];
-                // Update interaction: d_loss/d_interaction[i][j] = mask_features[i] * d_transformed[j]
-                self.interaction[i][j] -= lr * mask_features[i] * d_transformed[j];
-            }
-        }
-
-        (d_mask_features, d_rule_embed)
-    }
-
-    /// Backprop through mask MLP.
-    ///
-    /// Returns d_expr_embed and updates mask_mlp weights.
-    fn backprop_mask_mlp(
-        &mut self,
-        d_out: &[f32; EMBED_DIM],
-        expr_embed: &[f32; EMBED_DIM],
-        h_pre_relu: &[f32; MLP_HIDDEN],
-        lr: f32,
-    ) -> [f32; EMBED_DIM] {
-        // d_out → w2, b2
-        // d_h (post-ReLU) = d_out @ w2^T
-        let mut d_h = [0.0f32; MLP_HIDDEN];
-        for j in 0..MLP_HIDDEN {
-            for k in 0..EMBED_DIM {
-                d_h[j] += d_out[k] * self.mask_mlp_w2[j][k];
-                // Update w2
-                let h_relu = h_pre_relu[j].max(0.0);
-                self.mask_mlp_w2[j][k] -= lr * h_relu * d_out[k];
-            }
-        }
-
-        // Update b2
-        for k in 0..EMBED_DIM {
-            self.mask_mlp_b2[k] -= lr * d_out[k];
-        }
-
-        // ReLU backward
-        for j in 0..MLP_HIDDEN {
-            if h_pre_relu[j] <= 0.0 {
-                d_h[j] = 0.0;
-            }
-        }
-
-        // d_h → w1, b1, d_expr_embed
-        let mut d_expr_embed = [0.0f32; EMBED_DIM];
-        for i in 0..EMBED_DIM {
-            for j in 0..MLP_HIDDEN {
-                d_expr_embed[i] += d_h[j] * self.mask_mlp_w1[i][j];
-                // Update w1
-                self.mask_mlp_w1[i][j] -= lr * expr_embed[i] * d_h[j];
-            }
-        }
-
-        // Update b1
-        for j in 0..MLP_HIDDEN {
-            self.mask_mlp_b1[j] -= lr * d_h[j];
-        }
-
-        d_expr_embed
-    }
-
-    /// Backprop through rule MLP.
-    ///
-    /// Updates rule_mlp weights. Rule features are fixed, so no gradient returned.
-    fn backprop_rule_mlp(
-        &mut self,
-        d_out: &[f32; EMBED_DIM],
-        rule_features: &[f32; RULE_FEATURE_DIM],
-        h_pre_relu: &[f32; MLP_HIDDEN],
-        lr: f32,
-    ) {
-        // d_out → w2, b2
-        let mut d_h = [0.0f32; MLP_HIDDEN];
-        for j in 0..MLP_HIDDEN {
-            for k in 0..EMBED_DIM {
-                d_h[j] += d_out[k] * self.rule_mlp_w2[j][k];
-                let h_relu = h_pre_relu[j].max(0.0);
-                self.rule_mlp_w2[j][k] -= lr * h_relu * d_out[k];
-            }
-        }
-
-        // Update b2
-        for k in 0..EMBED_DIM {
-            self.rule_mlp_b2[k] -= lr * d_out[k];
-        }
-
-        // ReLU backward
-        for j in 0..MLP_HIDDEN {
-            if h_pre_relu[j] <= 0.0 {
-                d_h[j] = 0.0;
-            }
-        }
-
-        // d_h → w1, b1
-        for i in 0..RULE_FEATURE_DIM {
-            for j in 0..MLP_HIDDEN {
-                self.rule_mlp_w1[i][j] -= lr * rule_features[i] * d_h[j];
-            }
-        }
-
-        for j in 0..MLP_HIDDEN {
-            self.rule_mlp_b1[j] -= lr * d_h[j];
-        }
-    }
-
-    /// Backprop through value MLP.
-    ///
-    /// Returns d_expr_embed and updates value_mlp weights.
-    fn backprop_value_mlp(
-        &mut self,
-        d_cost: f32,
-        expr_embed: &[f32; EMBED_DIM],
-        h_pre_relu: &[f32; MLP_HIDDEN],
-        lr: f32,
-    ) -> [f32; EMBED_DIM] {
-        // d_cost → w2, b2
-        let mut d_h = [0.0f32; MLP_HIDDEN];
-        for j in 0..MLP_HIDDEN {
-            d_h[j] = d_cost * self.value_mlp_w2[j];
-            let h_relu = h_pre_relu[j].max(0.0);
-            self.value_mlp_w2[j] -= lr * h_relu * d_cost;
-        }
-
-        self.value_mlp_b2 -= lr * d_cost;
-
-        // ReLU backward
-        for j in 0..MLP_HIDDEN {
-            if h_pre_relu[j] <= 0.0 {
-                d_h[j] = 0.0;
-            }
-        }
-
-        // d_h → w1, b1, d_expr_embed
-        let mut d_expr_embed = [0.0f32; EMBED_DIM];
-        for i in 0..EMBED_DIM {
-            for j in 0..MLP_HIDDEN {
-                d_expr_embed[i] += d_h[j] * self.value_mlp_w1[i][j];
-                self.value_mlp_w1[i][j] -= lr * expr_embed[i] * d_h[j];
-            }
-        }
-
-        for j in 0..MLP_HIDDEN {
-            self.value_mlp_b1[j] -= lr * d_h[j];
-        }
-
-        d_expr_embed
-    }
-}
-
-/// Dot product of two arrays.
-#[inline]
-fn dot(a: &[f32; HIDDEN_DIM], b: &[f32; HIDDEN_DIM]) -> f32 {
-    let mut sum = 0.0;
-    for i in 0..HIDDEN_DIM {
-        sum += a[i] * b[i];
-    }
-    sum
-}
-
-/// Sigmoid activation.
-#[inline]
-fn sigmoid(x: f32) -> f32 {
-    1.0 / (1.0 + libm::expf(-x))
-}
-
-/// Softmax with temperature scaling.
-///
-/// `softmax(x_i / temp)` - higher temperature = more uniform distribution.
-#[must_use]
-fn softmax_with_temperature(logits: &[f32], temperature: f32) -> Vec<f32> {
-    if logits.is_empty() {
-        return Vec::new();
-    }
-
-    let temp = temperature.max(0.01); // Avoid division by zero
-
-    // Scale by temperature
-    let scaled: Vec<f32> = logits.iter().map(|&x| x / temp).collect();
-
-    // Numerical stability: subtract max
-    let max_val = scaled.iter().cloned().fold(f32::NEG_INFINITY, f32::max);
-    let exps: Vec<f32> = scaled.iter().map(|&x| libm::expf(x - max_val)).collect();
-    let sum: f32 = exps.iter().sum();
-
-    if sum < 1e-10 {
-        // Uniform fallback
-        vec![1.0 / logits.len() as f32; logits.len()]
-    } else {
-        exps.iter().map(|&e| e / sum).collect()
     }
 }
 
@@ -3388,193 +1739,18 @@ mod tests {
 
     #[test]
     fn consolidated_param_count() {
-        // Param count should include backbone + all unified heads
+        // Param count should include the whole backbone + value head.
         let count = ExprNnue::param_count();
-        // Backbone: embeddings + w1 + b1 = ~9,728
-        // Plus: expr_proj + value_mlp + mask_mlp + rule_mlp + rule_proj + interaction + bias
+        // Backbone (embeddings + w1 + b1 + trunk) ~13,808, plus expr_proj +
+        // value_mlp ~3,089 = 16,897 live params (67,588 bytes) — see
+        // `Self::save`'s doc for the split from the saturation head's
+        // (separately-checkpointed) weights.
         assert!(count > 10_000, "Should have >10k params, got {}", count);
         assert!(
             ExprNnue::memory_bytes() < 200_000,
             "NNUE should use < 200KB, got {} bytes",
             ExprNnue::memory_bytes()
         );
-    }
-
-    // ========================================================================
-    // Unified Mask Architecture Tests
-    // ========================================================================
-
-    #[test]
-    fn rule_features_initialization() {
-        let mut rule_features = RuleFeatures::new();
-
-        // All features should be zero initially
-        for r in 0..10 {
-            for f in rule_features.get(r) {
-                assert!(*f == 0.0, "Initial features should be zero");
-            }
-        }
-
-        // Set features for a rule
-        rule_features.set(0, [0.25, 0.3, 1.0, 1.0, 0.0, 1.0, 0.5, 1.0]);
-        let features = rule_features.get(0);
-        assert!((features[0] - 0.25).abs() < 1e-6, "Category should be set");
-        assert!(
-            (features[3] - 1.0).abs() < 1e-6,
-            "Commutative flag should be set"
-        );
-    }
-
-    #[test]
-    fn encode_rule_deterministic() {
-        let net = ExprNnue::new_random(42);
-        let features = [0.25, 0.3, 1.0, 1.0, 0.0, 1.0, 0.5, 1.0];
-
-        let embed1 = net.encode_rule(&features);
-        let embed2 = net.encode_rule(&features);
-
-        // Same input should produce same output
-        for i in 0..EMBED_DIM {
-            assert!(
-                (embed1[i] - embed2[i]).abs() < 1e-6,
-                "encode_rule should be deterministic at dim {}",
-                i
-            );
-        }
-
-        // Embedding should be finite
-        for i in 0..EMBED_DIM {
-            assert!(
-                embed1[i].is_finite(),
-                "Rule embedding should be finite at dim {}",
-                i
-            );
-        }
-    }
-
-    #[test]
-    fn verify_encode_all_rules() {
-        let net = ExprNnue::new_random(42);
-        let mut rule_features = RuleFeatures::new();
-
-        // Set up a few rules with different features
-        rule_features.set(0, [0.0, 0.2, 0.0, 1.0, 0.0, 0.0, 0.1, 0.0]); // algebraic
-        rule_features.set(1, [0.25, 0.5, -1.0, 0.0, 1.0, 1.0, 0.3, 0.0]); // peephole
-        rule_features.set(2, [0.75, 0.8, 1.0, 0.0, 0.0, 0.0, 0.05, 1.0]); // cross-cutting
-
-        let embeds = net.encode_all_rules(&rule_features, 3);
-
-        assert_eq!(embeds.len(), 3, "Should encode exactly 3 rules");
-
-        // Each embedding should be finite
-        for (r, embed) in embeds.iter().enumerate() {
-            for (d, &val) in embed.iter().enumerate() {
-                assert!(val.is_finite(), "Rule {} dim {} should be finite", r, d);
-            }
-        }
-
-        // Different features should produce different embeddings
-        let diff_01: f32 = embeds[0]
-            .iter()
-            .zip(embeds[1].iter())
-            .map(|(a, b)| (a - b).abs())
-            .sum();
-        let diff_02: f32 = embeds[0]
-            .iter()
-            .zip(embeds[2].iter())
-            .map(|(a, b)| (a - b).abs())
-            .sum();
-
-        assert!(
-            diff_01 > 1e-3,
-            "Different rules should have different embeddings"
-        );
-        assert!(
-            diff_02 > 1e-3,
-            "Different rules should have different embeddings"
-        );
-    }
-
-    #[test]
-    fn bilinear_score_computation() {
-        let net = ExprNnue::new_random(42);
-
-        // Create test vectors
-        let mask_features = [1.0f32; EMBED_DIM];
-        let rule_embed = [1.0f32; EMBED_DIM];
-
-        let score = net.bilinear_score(&mask_features, &rule_embed);
-
-        assert!(score.is_finite(), "Bilinear score should be finite");
-
-        // Manual verification: score = dot(mask @ interaction + bias_proj, rule)
-        // With all-ones vectors: sum of interaction matrix + sum of bias_proj
-        let mut expected = 0.0f32;
-        for i in 0..EMBED_DIM {
-            for j in 0..EMBED_DIM {
-                expected += net.interaction[i][j];
-            }
-        }
-        for k in 0..EMBED_DIM {
-            expected += net.mask_bias_proj[k];
-        }
-        assert!(
-            (score - expected).abs() < 1e-4,
-            "Bilinear computation mismatch: got {}, expected {}",
-            score,
-            expected
-        );
-    }
-
-    #[test]
-    fn verify_randomize_mask_only() {
-        let mut net = ExprNnue::new();
-
-        // Set some backbone values that should be preserved
-        net.embeddings.e[OpKind::Var][0] = 1.234;
-        net.w1[0][0] = 5.678;
-        net.b1[0] = 0.999;
-        net.expr_proj_w[0][0] = 2.345; // shared projection - should be preserved
-
-        // Initially mask-specific weights should be zero
-        let initial_mask_sum: f32 = net.mask_mlp_w1.iter().flatten().map(|x| x.abs()).sum();
-        assert!(
-            initial_mask_sum < 1e-6,
-            "Initial mask weights should be zero"
-        );
-
-        // Randomize mask-only
-        net.randomize_mask_only(42);
-
-        // Backbone should be PRESERVED
-        assert!(
-            (net.embeddings.e[OpKind::Var][0] - 1.234).abs() < 1e-6,
-            "Embeddings should be preserved"
-        );
-        assert!(
-            (net.w1[0][0] - 5.678).abs() < 1e-6,
-            "w1 should be preserved"
-        );
-        assert!((net.b1[0] - 0.999).abs() < 1e-6, "b1 should be preserved");
-        assert!(
-            (net.expr_proj_w[0][0] - 2.345).abs() < 1e-6,
-            "expr_proj should be preserved"
-        );
-
-        // Mask-specific weights should now be non-zero
-        let final_mask_sum: f32 = net.mask_mlp_w1.iter().flatten().map(|x| x.abs()).sum();
-        assert!(
-            final_mask_sum > 1.0,
-            "Randomized mask weights should be non-zero"
-        );
-
-        // Interaction matrix should be near identity diagonal
-        for i in 0..EMBED_DIM {
-            assert!(
-                (net.interaction[i][i] - 1.0).abs() < 0.5,
-                "Diagonal of interaction should be near 1.0"
-            );
-        }
     }
 
     // ========================================================================
@@ -3606,177 +1782,5 @@ mod tests {
             );
         }
         assert_eq!(acc.edge_count, 0);
-    }
-
-    // ========================================================================
-    // structural_hash and from_expr_dedup tests
-    // ========================================================================
-
-    // ========================================================================
-    // GraphAccumulator normalization tests
-    // ========================================================================
-
-    #[test]
-    fn graph_acc_normalize_unit_norm_per_section() {
-        let emb = OpEmbeddings::new_random(42);
-        let mut gacc = GraphAccumulator::new();
-        // Build a non-trivial accumulator: several edges
-        gacc.add_edge(&emb, OpKind::Add, OpKind::Mul);
-        gacc.add_edge(&emb, OpKind::Mul, OpKind::Var);
-        gacc.add_edge(&emb, OpKind::Mul, OpKind::Var);
-        gacc.add_edge(&emb, OpKind::Add, OpKind::Var);
-
-        let normed = gacc.normalized();
-
-        // Each section should have L2 norm ~1.0
-        let section_norm = |start: usize, end: usize| -> f32 {
-            let sum_sq: f32 = normed.values[start..end].iter().map(|v| v * v).sum();
-            sqrtf(sum_sq)
-        };
-        let eps = 1e-5;
-        assert!(
-            (section_norm(0, K) - 1.0).abs() < eps,
-            "parent section norm should be 1.0"
-        );
-        assert!(
-            (section_norm(K, 2 * K) - 1.0).abs() < eps,
-            "child section norm should be 1.0"
-        );
-        assert!(
-            (section_norm(2 * K, 3 * K) - 1.0).abs() < eps,
-            "1-hop VSA section norm should be 1.0"
-        );
-        // 2-hop section is zero (no 2-hop edges added) — normalization should leave it as zeros.
-        let hop2_norm = section_norm(3 * K, 4 * K);
-        assert!(
-            hop2_norm < eps,
-            "2-hop VSA section should be zero when no 2-hop edges added, got {hop2_norm}"
-        );
-    }
-
-    #[test]
-    fn graph_acc_normalize_scalars_preserved() {
-        let emb = OpEmbeddings::new_random(42);
-        let mut gacc = GraphAccumulator::new();
-        gacc.add_edge(&emb, OpKind::Add, OpKind::Mul);
-        gacc.add_leaf();
-        gacc.add_leaf();
-        gacc.node_budget = 100;
-        gacc.epoch_budget = 50;
-
-        let normed = gacc.normalized();
-        assert_eq!(
-            normed.edge_count, gacc.edge_count,
-            "edge_count must be preserved"
-        );
-        assert_eq!(
-            normed.node_count, gacc.node_count,
-            "node_count must be preserved"
-        );
-        assert_eq!(
-            normed.node_budget, gacc.node_budget,
-            "node_budget must be preserved"
-        );
-        assert_eq!(
-            normed.epoch_budget, gacc.epoch_budget,
-            "epoch_budget must be preserved"
-        );
-    }
-
-    #[test]
-    fn graph_acc_normalize_zero_is_safe() {
-        // A fresh (zero) accumulator should normalize without NaN/Inf.
-        let gacc = GraphAccumulator::new();
-        let normed = gacc.normalized();
-        for (i, &v) in normed.values.iter().enumerate() {
-            assert!(v.is_finite(), "values[{i}] must be finite, got {v}");
-            assert_eq!(
-                v, 0.0,
-                "zero accumulator must stay zero after normalization"
-            );
-        }
-    }
-
-    #[test]
-    fn graph_acc_normalize_in_place_matches_normalized() {
-        let emb = OpEmbeddings::new_random(42);
-        let mut gacc = GraphAccumulator::new();
-        gacc.add_edge(&emb, OpKind::Add, OpKind::Mul);
-        gacc.add_edge(&emb, OpKind::Mul, OpKind::Var);
-
-        let copy_normed = gacc.normalized();
-
-        let mut in_place = gacc.clone();
-        in_place.normalize_in_place();
-
-        for i in 0..GRAPH_ACC_DIM {
-            assert!(
-                (copy_normed.values[i] - in_place.values[i]).abs() < 1e-9,
-                "values[{i}] mismatch: normalized()={} vs normalize_in_place()={}",
-                copy_normed.values[i],
-                in_place.values[i]
-            );
-        }
-    }
-
-    #[test]
-    fn graph_acc_normalize_scale_invariance() {
-        // Doubling all edges (adding each edge twice) should yield the same
-        // normalized vector, proving scale invariance.
-        let emb = OpEmbeddings::new_random(42);
-
-        let mut small = GraphAccumulator::new();
-        small.add_edge(&emb, OpKind::Add, OpKind::Mul);
-        small.add_edge(&emb, OpKind::Mul, OpKind::Var);
-
-        let mut large = GraphAccumulator::new();
-        for _ in 0..10 {
-            large.add_edge(&emb, OpKind::Add, OpKind::Mul);
-            large.add_edge(&emb, OpKind::Mul, OpKind::Var);
-        }
-
-        let small_n = small.normalized();
-        let large_n = large.normalized();
-
-        for i in 0..GRAPH_ACC_DIM {
-            assert!(
-                (small_n.values[i] - large_n.values[i]).abs() < 1e-5,
-                "normalized vectors should match regardless of scale: values[{i}] small={} large={}",
-                small_n.values[i],
-                large_n.values[i]
-            );
-        }
-    }
-
-    #[test]
-    fn graph_acc_normalize_idempotent() {
-        // Normalizing twice should produce the same result.
-        let emb = OpEmbeddings::new_random(42);
-        let mut gacc = GraphAccumulator::new();
-        gacc.add_edge(&emb, OpKind::Add, OpKind::Mul);
-        gacc.add_edge(&emb, OpKind::Mul, OpKind::Var);
-
-        let once = gacc.normalized();
-        let twice = once.normalized();
-
-        for i in 0..GRAPH_ACC_DIM {
-            assert!(
-                (once.values[i] - twice.values[i]).abs() < 1e-6,
-                "normalization must be idempotent: values[{i}] once={} twice={}",
-                once.values[i],
-                twice.values[i]
-            );
-        }
-    }
-
-    // ========================================================================
-    // GraphAccumulator incremental remove tests
-    // ========================================================================
-
-    #[test]
-    fn graph_acc_remove_leaf_saturates_at_zero() {
-        let mut acc = GraphAccumulator::new();
-        acc.remove_leaf();
-        assert_eq!(acc.node_count, 0, "node_count must not underflow");
     }
 }
